% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{AddAttributes}
\alias{AddAttributes}
\title{Add attributes to proposed.reconstruction to be same as spectra}
\usage{
AddAttributes(proposed.reconstruction, spectra)
}
\description{
Add attributes to proposed.reconstruction to be same as spectra
}
\keyword{internal}
