% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureProportions.R
\name{AddMinusEFor192}
\alias{AddMinusEFor192}
\title{Translate SBS96 signature IDs to SBS192 signature IDs.}
\usage{
AddMinusEFor192(sig.ids)
}
\arguments{
\item{sig.ids}{Character vector of SBS96 signature IDs.}
}
\value{
Character vector of corresponding SBS192 signature IDs; some
  have "-E" (for exmome) post-pended.
}
\description{
Translate SBS96 signature IDs to SBS192 signature IDs.
}
\keyword{internal}
