% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSigActivity.R
\name{AddSigActivity1}
\alias{AddSigActivity1}
\title{Add contributing signature activity information for one sample}
\usage{
AddSigActivity1(
  spect,
  exposure,
  sigs,
  sigs.presence.prop,
  nbinom.size = 5,
  likelihood.dist = "multinom",
  use.sparse.assign = FALSE
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs.}

\item{sigs}{The signatures with which we are trying to reconstruct \code{spect}.
A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog. The column
names of \code{sigs} should be a superset of row names of \code{exposure}.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{nbinom.size}{The dispersion parameter for the negative binomial
distribution; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}.}

\item{likelihood.dist}{The probability distribution used to calculate the
likelihood, can be either "multinom" (multinomial distribution) or
"neg.binom" (negative binomial distribution).}

\item{use.sparse.assign}{Whether to use sparse assignment. If \code{TRUE},
arguments designed for Maximum A Posteriori assignment such as
\code{sigs.presence.prop} will be ignored.}
}
\value{
A list of elements:
\itemize{
\item \code{original.spect}: The original \code{spect} with total mutation counts
added to its column name. An additional attribute "exposure" from
\code{exposure} is also added.
\item \code{reconstructed.spect}: The reconstructed spectrum using \code{sigs}
and \code{exposure}. Its column name has the total mutation counts and cosine
similarity with the original \code{spect}.
\item \code{contributing.sigs}: The contributing signatures to the original
\code{spect}. The column names of each contributing signature has mutation
counts attributed to this signature, its contribution proportion and proposed
etiology. (If the etiology is unknown, then will be blank.)
\item \code{distances}: Various distances and similarities between the original
spectrum and \code{reconstructed.spect}.
}
}
\description{
Add contributing signature activity information for one sample
}
\details{
This function calls \code{\link{ReconstructSpectrum}},
\code{\link{LLHSpectrumNegBinom}} and \code{\link{LLHSpectrumMAP}}.
}
\section{Remark}{
 The column names of \code{spect} should be the same as the
column name of \code{exposure}.
}

\keyword{internal}
