% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnySigSubsetPresent.R
\name{AnySigSubsetPresentInternal}
\alias{AnySigSubsetPresentInternal}
\title{For each combination of several signatures, determine if the combination is plausibly needed to reconstruct a spectrum.}
\usage{
AnySigSubsetPresentInternal(
  spect,
  all.sigs,
  Ha.sigs.indices,
  m.opts,
  max.mc.cores
)
}
\arguments{
\item{spect}{The spectrum to be reconstructed, as single column matrix or
\code{\link[ICAMS]{ICAMS}} catalog.}

\item{all.sigs}{The matrix or catalog of all signatures of possible interest,
which consist of the signatures for \eqn{H_0} and for the alternative
hypotheses.}

\item{Ha.sigs.indices}{An integer vector of the indices of the signatures
that are in the various \eqn{H_a}'s.}

\item{m.opts}{Controls the numerical search for maximum likelihood
reconstructions of \code{spect} plus some additional
flags; see \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
If \code{NULL} defaults to \eqn{2^{n_a} - 1},
 where \eqn{n_a} is the length of \code{Ha.sigs.indices} -- except on
 MS Windows machines, where it defaults to 1.}
}
\value{
A list with two elements: \describe{

\item{\code{H0.info}}{contains the sub-elements \describe{
   \item{\code{loglh}}{The log likelihood associated with \eqn{H_0}.}
   \item{\code{exposure}}{The signature attributions (exposures) corresponding
        to the \eqn{H_0} log likelihood.}
   \item{\code{everything.else}}{A sub-list with information on the output
        of the numerical optimization that provided \code{loglh}.}
}}

\item{\code{all.Ha.info}}{A list with one sub-element for each non-empty subset of
  \code{Ha.sigs.indices}. Each sub-element is a list with elements that include \describe{

  \item{\code{sigs.added}}{The identifiers of the (additional) signatures
       tested.}

  \item{\code{p}}{The \eqn{p} value for the likelihood-ratio test.} This
  this \eqn{p} value can be \code{NaN} when the likelihoods of (\eqn{H_0} and \eqn{H_a})
  are both \code{-Inf}. This can occur if there are are mutation classes in the spectra
  that are > 0 but that have 0 probability in all the available input signatures.
  This is unlikely to occur, since most spectra have non-0 (albeit very small)
  probabilities for most mutation classes.

  \item{\code{df}}{The degrees of freedom of the likelihood-ratio test
     (equal to the number of signatures in \code{sigs.added}).}
     }
}}
}
\description{
Please see \strong{Details}.
}
\details{
Let \eqn{H_0} be the likelihood that
the signatures specified by \code{all.sigs[, -Ha.sigs.indicies, drop = FALSE]}
generated the observed spectrum, \code{spect}.
For each non-empty subset, \eqn{S},
of \code{Ha.sigs.indices} let \eqn{H_a}
be the likelihood that all the signatures in \eqn{H_0}
plus the signatures specified by \eqn{S} generated \code{spect}.
Return a list with the results of likelihood ratio tests of
all \eqn{H_a}'s against \eqn{H_0}.
}
\keyword{internal}
