% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BG.MEG.kucab.bg.test}
\alias{BG.MEG.kucab.bg.test}
\title{List of test data for background subtraction based on Kucab 2016 Methyleugenol.}
\format{An object of class \code{list} of length 6.}
\usage{
BG.MEG.kucab.bg.test
}
\description{
Generated by \code{data-raw/prepare.BG.MEG.kucab.test.data.R}.
}
\keyword{datasets}
