% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BG.MEG.Test}
\alias{BG.MEG.Test}
\title{List of test data for background subtraction based on Kucab 2016 Methyleugenol.}
\format{An object of class \code{list} of length 6.}
\usage{
BG.MEG.Test
}
\description{
List of test data for background subtraction based on Kucab 2016 Methyleugenol.
}
\keyword{datasets}
