% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BTSG.WGS.PCAWG}
\alias{BTSG.WGS.PCAWG}
\title{Demonstration data: an \code{\link[ICAMS]{ICAMS}} catalog.}
\format{An \code{\link[ICAMS]{ICAMS}} catalog with 
mutations from 35 biliary tract tumors.}
\source{
Cite Alexandrov et al, 2019
}
\usage{
BTSG.WGS.PCAWG
}
\description{
Demonstration data: an \code{\link[ICAMS]{ICAMS}} catalog.
}
\keyword{datasets}
