% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcuateDistance.R
\name{CalculateDistance}
\alias{CalculateDistance}
\title{Calculate various distances information between spectra and reconstructed spectra}
\usage{
CalculateDistance(
  spectra,
  exposure,
  sigs,
  sigs.presence.prop,
  nbinom.size = 5,
  likelihood.dist = "multinom",
  use.sparse.assign = TRUE
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs.}

\item{sigs}{The signatures with which we are trying to reconstruct \code{spectra}.
A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog. The column
names of \code{sigs} should be a superset of row names of \code{exposure}.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{nbinom.size}{The dispersion parameter for the negative binomial
distribution; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}.}

\item{likelihood.dist}{The probability distribution used to calculate the
likelihood, can be either "multinom" (multinomial distribution) or
"neg.binom" (negative binomial distribution).}

\item{use.sparse.assign}{Whether to use sparse assignment. If \code{TRUE},
arguments designed for Maximum A Posteriori assignment such as
\code{sigs.presence.prop} will be ignored.}
}
\value{
A data frame with row names being the sample names in \code{spectra}.
The columns show various distances information between \code{spectra} and reconstructed
spectra using \code{sigs} and \code{exposure}. The column names are:
\itemize{
\item log.likelihood
\item euclidean
\item manhattan
\item cosine
\item mutations
\item scaled.manhattan
}
}
\description{
Calculate various distances information between spectra and reconstructed spectra
}
\keyword{internal}
