% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisplatin.exposed.HepG2.96}
\alias{cisplatin.exposed.HepG2.96}
\title{Test experimental spectra}
\format{XXXX}
\usage{
cisplatin.exposed.HepG2.96
}
\description{
Test experimental spectra
}
\examples{
# XXXX
}
\keyword{datasets}
