% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyAttributes.R
\name{CopyAttributes}
\alias{CopyAttributes}
\title{If from is an ICAMS catalog copy its attributes to to to, otherwise just make to a counts catalog}
\usage{
CopyAttributes(from, to)
}
\description{
If from is an ICAMS catalog copy its attributes to to to, otherwise just make to a counts catalog
}
\keyword{internal}
