% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cossim}
\alias{cossim}
\title{Cosine similarity with more forgiving argment types.}
\usage{
cossim(v1, v2)
}
\arguments{
\item{v1}{A vector or single-column matrix}

\item{v2}{A vector or single-column matrix}
}
\description{
Cosine similarity with more forgiving argment types.
}
