% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{DistanceMeasures}
\alias{DistanceMeasures}
\title{Calculate several distance measures between a spectrum and its reconstruction}
\usage{
DistanceMeasures(spect, recon, nbinom.size, model, sigs.presence.prop)
}
\description{
Calculate several distance measures between a spectrum and its reconstruction
}
\keyword{internal}
