% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseAssignActivity1.R
\name{DistanceMeasuresSparse}
\alias{DistanceMeasuresSparse}
\title{Calculate several distance measures between a spectrum and its reconstruction using
sparse assignment}
\usage{
DistanceMeasuresSparse(
  spect,
  recon,
  nbinom.size,
  likelihood.dist = "multinom",
  signatures = NULL,
  m.opts = NULL
)
}
\arguments{
\item{spect}{The spectrum we are trying to reconstruct.}

\item{recon}{The unrounded reconstruction.}

\item{nbinom.size}{\strong{Only} needed when \code{likelihood.dist =
"neg.binom"}.The dispersion parameter for the negative binomial
distribution; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}.}

\item{likelihood.dist}{The probability distribution used to calculate the
likelihood, can be either "multinom" (multinomial distribution) or
"neg.binom" (negative binomial distribution).}

\item{signatures}{\strong{Only} used to compute distances for quadratic
programming assignment. Signature as a matrix or data frame, with each row
one mutation type (g.e. CCT > CAT or CC > TT) and each column a signature.
It should be the proposed signatures used by sparse
assignment to reconstruct \code{spect}. It is needed to calculate distances
of quadratic programming assignment.}
}
\value{
A data frame whose first column indicates the distance method. The
  second column \code{proposed.assignment} shows the values of various
  distances using sparse assignment. 
  
  When \code{signatures} is \strong{not} NULL, there will be a third column
  \code{QP.assignment} shows the values of various distances using quadratic
  programming assignment.
}
\description{
The idea is to provide several measures of how well the 
reconstruction matches the spectrum.
}
\keyword{internal}
