% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{EDist2SpectRounded}
\alias{EDist2SpectRounded}
\title{Euclidean reconstruction error.}
\usage{
EDist2SpectRounded(exp, sig.names, spect)
}
\description{
Euclidean reconstruction error.
}
\keyword{internal}
