% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateBackground.R
\name{FindSigMinusBGOpt}
\alias{FindSigMinusBGOpt}
\title{Return a default value to pass as the \code{m.opts} argument to \code{\link{SeparateSignatureFromBackground}}.}
\usage{
FindSigMinusBGOpt()
}
\description{
Return a default value to pass as the \code{m.opts} argument to \code{\link{SeparateSignatureFromBackground}}.
}
