% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{FindSignatureMinusBackground}
\alias{FindSignatureMinusBackground}
\title{Estimate a signature from experimentally exposed spectra minus a background signature.}
\usage{
FindSignatureMinusBackground(spectra, bg.sig.info,
  algorithm = "NLOPT_LN_COBYLA", maxeval = 1000, print_level = 0,
  xtol_rel = 0.001, xtol_abs = 1e-04, start.b.fraction = 0.1)
}
\description{
Let the input spectra be s1, s2, ...
}
\details{
Find target.signature that maximize the likelihoods:

```
max_(b1, target.signature){s1 | 
            b1, target.signature, background.sig, 
            background.sig.nbinom.size, background.sig.count.mu,
     background.sig.count.nbinom.size}
b1 * background.sig + (total-mut1 - b1) * target.sig * prob(b1), \cr
b2 * background.sig + (total-mut2 - b2) * target.sig * prob(b2), \cr
... 
````

See \code{\link{ObjFn1}}.
}
\keyword{internal}
