% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_ineq_for_ObjFnBinomMaxLH2.R
\name{g_ineq_for_ObjFnBinomMaxLH2}
\alias{g_ineq_for_ObjFnBinomMaxLH2}
\title{Function to constrain the sum of estimated exposures to the number of mutations in the spectrum.}
\usage{
g_ineq_for_ObjFnBinomMaxLH2(exp, spectrum, sigs, nbinom.size)
}
\arguments{
\item{exp}{A numeric vector of exposures.}

\item{spectrum}{The observed spectrum we are trying to reconstruct.}

\item{sigs}{The signatures with which we are trying to reconstruct
the spectrum. (Ignored in this function but used by \code{\link[nloptr]{nloptr}}.)}

\item{nbinom.size}{Dispersion parameter.
(Ignored in this function but used by \code{\link[nloptr]{nloptr}}.)}
}
\description{
See \code{\link[nloptr]{nloptr}} to understand how this function is used.
}
\keyword{internal}
