% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_ineq_for_ObjFnDMMaxLH.R
\name{g_ineq_for_ObjFnDMMaxLH}
\alias{g_ineq_for_ObjFnDMMaxLH}
\title{Function to constrain the sum of estimated exposures to the number of
mutations in the spectrum.}
\usage{
g_ineq_for_ObjFnDMMaxLH(exp, spectrum, sigs)
}
\arguments{
\item{exp}{A numeric vector of exposures.}

\item{spectrum}{The observed spectrum we are trying to reconstruct.}

\item{sigs}{The signatures with which we are trying to reconstruct
the spectrum. (Ignored in this function but used by \code{\link[nloptr]{nloptr}}.)}
}
\description{
See \code{\link[nloptr]{nloptr}} to understand how this function is used.
}
\keyword{internal}
