% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetAllTestedTables}
\alias{GetAllTestedTables}
\title{Retrieve all tested tables from the output generated by running
\code{MAPAssignActivity1} on multiple samples}
\usage{
GetAllTestedTables(list.of.MAP.out)
}
\description{
Retrieve all tested tables from the output generated by running
\code{MAPAssignActivity1} on multiple samples
}
\keyword{internal}
