% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{GetAltSolutions}
\alias{GetAltSolutions}
\title{Get alternative solutions that are statistically as good as the best solution
from \code{MAPAssignActivitiy1}}
\usage{
GetAltSolutions(
  tibble,
  spectrum,
  sigs,
  mc.cores = 1,
  sparse.assign = FALSE,
  wt.thresh = 0.95,
  p.thresh = 0.05
)
}
\description{
Get alternative solutions that are statistically as good as the best solution
from \code{MAPAssignActivitiy1}
}
\keyword{internal}
