% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetAltSolutionsTables}
\alias{GetAltSolutionsTables}
\title{Retrieve alternative solutions tables from the output generated by running
\code{MAPAssignActivity1} on multiple samples}
\usage{
GetAltSolutionsTables(list.of.MAP.out)
}
\description{
Retrieve alternative solutions tables from the output generated by running
\code{MAPAssignActivity1} on multiple samples
}
\keyword{internal}
