% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetExposureInfo}
\alias{GetExposureInfo}
\title{Retrieve exposure information from the output generated by running
\code{MAPAssignActivity1} on multiple samples}
\usage{
GetExposureInfo(list.of.MAP.out, type.of.exposure = "MAP")
}
\arguments{
\item{list.of.MAP.out}{Output generated by \code{MAPAssignActivity}.}

\item{type.of.exposure}{Type of exposure information to retrieve, either
"MAP" or "best.sparse".}
}
\description{
Retrieve exposure information from the output generated by running
\code{MAPAssignActivity1} on multiple samples
}
\keyword{internal}
