% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetReconstructionInfo}
\alias{GetReconstructionInfo}
\title{Retrieve reconstruction information from the output generated by running
\code{MAPAssignActivity1} on multiple samples}
\usage{
GetReconstructionInfo(list.of.MAP.out)
}
\description{
Retrieve reconstruction information from the output generated by running
\code{MAPAssignActivity1} on multiple samples
}
\keyword{internal}
