% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetTimeForMAPAssign}
\alias{GetTimeForMAPAssign}
\title{Retrieve time.for.MAP.assign from the output generated by running
\code{MAPAssignActivity1} on multiple samples}
\usage{
GetTimeForMAPAssign(list.of.MAP.out)
}
\description{
Retrieve time.for.MAP.assign from the output generated by running
\code{MAPAssignActivity1} on multiple samples
}
\keyword{internal}
