% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HepG2.background.spectra}
\alias{HepG2.background.spectra}
\title{Background spectra for HepG2.}
\format{An \code{[ICAMS]} \code{counts} catalog.}
\usage{
HepG2.background.spectra
}
\description{
Background spectra for HepG2.
}
\keyword{datasets}
