% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HepG2.bg.tests.no.noise}
\alias{HepG2.bg.tests.no.noise}
\title{Tests of HepG2 background subtraction}
\format{A data table}
\usage{
HepG2.bg.tests.no.noise
}
\description{
Tests of HepG2 background subtraction
}
\examples{

HepG2.bg.tests.no.noise[1:2, 1:10]
}
\keyword{datasets}
