% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.bg}
\alias{kucab.bg}
\title{A deprecated version of \code{\link{kucab.control.bg}}}
\format{An object of class \code{list} of length 5.}
\usage{
kucab.bg
}
\description{
A deprecated version of \code{\link{kucab.control.bg}}
}
\keyword{datasets}
