% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.control.bg}
\alias{kucab.control.bg}
\title{Background information from the controls in Kucab et al., 2019.}
\format{An object of class \code{list} of length 4.}
\usage{
kucab.control.bg
}
\description{
Estimated from the control spectra in \code{\link{kucab.spectra}}.
}
\keyword{datasets}
