% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.control.dist}
\alias{kucab.control.dist}
\title{Resampled mutations counts for combinations of 2:4 \code{kucab.controls}.}
\format{A list with elements 2:4, each of which is a vector of 
10,000 elements, each of which is the mean of total 
mutation counts of 2 to 4 control spectra.}
\usage{
kucab.control.dist
}
\description{
Resampled mutations counts for combinations of 2:4 \code{kucab.controls}.
}
\keyword{datasets}
