% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.control.spectra.ICAMS}
\alias{kucab.control.spectra.ICAMS}
\title{Control spectra from Kucab et al., 2019}
\format{An \code{\link[ICAMS]{ICAMS}} \code{counts} catalog with 35 samples.}
\usage{
kucab.control.spectra.ICAMS
}
\description{
Control spectra from Kucab et al., 2019
}
\keyword{datasets}
