% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.control.spectra.kucab}
\alias{kucab.control.spectra.kucab}
\title{The 35 control spectra using the Kucab row order.}
\format{An object of class \code{data.frame} with 96 rows and 35 columns.}
\usage{
kucab.control.spectra.kucab
}
\description{
The 35 control spectra using the Kucab row order.
}
\keyword{datasets}
