% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.controls}
\alias{kucab.controls}
\title{Control spectra from Kucab et al, 2019}
\format{An ICAMS catalog with 35 samples.}
\usage{
kucab.controls
}
\description{
Control spectra from Kucab et al, 2019
}
\keyword{datasets}
