% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.muts.control}
\alias{kucab.muts.control}
\title{Information on the control samples from Kucab et al., 2019}
\format{An object of class \code{data.frame} with 35 rows and 9 columns.}
\usage{
kucab.muts.control
}
\description{
Information on the control samples from Kucab et al., 2019
}
\keyword{datasets}
