% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.sigs.orig.row.order}
\alias{kucab.sigs.orig.row.order}
\title{Kucab et al., 2019 signatures in the row order as found on}
\format{An object of class \code{data.frame} with 96 rows and 55 columns.}
\source{
\url{https://data.mendeley.com/datasets/m7r4msjb4c/2#folder-9f2942fa-935b-4ac3-aa26-35784fb68a85}

The row labels are in column 1 in the order

A[C>A]A A[C>A]C A[C>A]G A[C>A]T C[C>A]A ...
}
\usage{
kucab.sigs.orig.row.order
}
\description{
Kucab et al., 2019 signatures in the row order as found on
}
\keyword{datasets}
