% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.spectra.ICAMS}
\alias{kucab.spectra.ICAMS}
\title{Kucab et al., 2019 spectra as a standard \code{ICAMS} count catalog.}
\format{An object of class \code{SBS96Catalog} (inherits from \code{matrix}) with 96 rows and 324 columns.}
\source{
https://github.com/xqzou/Cell_MutagenSig
}
\usage{
kucab.spectra.ICAMS
}
\description{
Kucab et al., 2019 spectra as a standard \code{ICAMS} count catalog.
}
\keyword{datasets}
