% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackgroundSubtractionSimpleTests.R
\name{KucabToICAMSSpectra}
\alias{KucabToICAMSSpectra}
\title{Convert a Kucab et al., spectra catalog to an ICAMS catalog.}
\usage{
KucabToICAMSSpectra(m)
}
\arguments{
\item{m}{The spectra (or signature) as a matrix, with
rownames e.g. "A[C>A]A", "A[C>G]C", ...}
}
\description{
The Zou code in Kucab et al., 2019 uses an unusual row order
and the first column has the mutation names as e.g.
A[C>T]G.
}
