% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{LLHOfSignatureGivenSpectrum}
\alias{LLHOfSignatureGivenSpectrum}
\title{Part of the objective function for one background-only spectrum.}
\usage{
LLHOfSignatureGivenSpectrum(spectrum, signature, nbinom.size)
}
\arguments{
\item{spectrum}{A single background spectrum}

\item{signature}{A signature as a numeric vector}

\item{nbinom.size}{The \code{size} argument for
\code{\link[stats]{dnbinom}}.}
}
\description{
Return the log likelihood of the observed spectrum
given the signature, and negative binomial dispersion
parameter \code{nbinom.size}.
}
\keyword{internal}
