% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLHSpectrumDM.R
\name{LLHSpectrumDM}
\alias{LLHSpectrumDM}
\title{Likelihood that 1 observed spectrum was generated from a vector of expected
mutation counts using Dirichlet-multinomial distribution}
\usage{
LLHSpectrumDM(spectrum, expected.counts, cp.factor = 1000, verbose = FALSE)
}
\arguments{
\item{spectrum}{An observed spectrum (a numeric vector)}

\item{expected.counts}{A vector of expected mutation counts, one expected
count for each mutation type. We want to know the likelihood that this
model generated the observed spectrum, assuming each mutational type
generates counts according to a Dirichlet-multinomial distribution with the
given \code{expected.counts} (argument \code{alpha} to
\code{\link[MGLM]{rdirmn}}).}

\item{cp.factor}{Concentration parameter factor. When calculating the
dirichlet multinomial likelihood, the concentration parameters \code{alpha}
is calculated as follows: 
\code{cp.factor} * \code{expected.counts} / \code{sum(expected.counts)}}

\item{verbose}{If \code{TRUE} print messages under some circumstances.}
}
\value{
\code{log(likelihood(spectrum | expected.counts))}, or, in more
  detail, the Dirichlet-multinomial likelihood that each element of the
  spectrum (i.e., the count for each mutation type e.g. ACT > AAT) was
  generated from the expected count for that mutation type using
  Dirichlet-multinomial distribution.
}
\description{
Likelihood that 1 observed spectrum was generated from a vector of expected
mutation counts using Dirichlet-multinomial distribution
}
