% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLHSpectrumMultinom.R
\name{LLHSpectrumMultinom}
\alias{LLHSpectrumMultinom}
\title{Likelihood that 1 observed spectrum was generated from a vector of expected
mutation counts using multinomial distribution}
\usage{
LLHSpectrumMultinom(spectrum, expected.counts, verbose = FALSE)
}
\arguments{
\item{spectrum}{An observed spectrum (a numeric vector)}

\item{expected.counts}{A vector of expected mutation counts, one expected
count for each mutation type. We want to know the likelihood that this
model generated the observed spectrum, assuming each mutational type
generates counts according to a multinomial distribution with the given
\code{expected.counts} (argument \code{prob} to
\code{\link[stats]{Multinom}}).}

\item{verbose}{If \code{TRUE} print messages under some circumstances.}
}
\value{
\code{log(likelihood(spectrum | expected.counts))}, or, in more
  detail, the multinomial likelihood that each element of the spectrum (i.e.,
  the count for each mutation type e.g. ACT > AAT) was generated from the
  expected count for that mutation type using multinomial distribution.
}
\description{
Likelihood that 1 observed spectrum was generated from a vector of expected
mutation counts using multinomial distribution
}
\keyword{internal}
