% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLHSpectrumNegBinom.R
\name{LLHSpectrumNegBinomDebug}
\alias{LLHSpectrumNegBinomDebug}
\title{A verbose version of \code{\link{LLHSpectrumNegBinom}} for testing}
\usage{
LLHSpectrumNegBinomDebug(
  spectrum,
  expected.counts,
  nbinom.size,
  verbose = FALSE
)
}
\arguments{
\item{spectrum}{An observed spectrum (a numeric vector)}

\item{expected.counts}{A vector of (integer) expected mutation counts, one
expected count for each mutation type. We want to know the
likelihood that this model generated the observed
spectrum, assuming each mutational types generates counts according to
a negative binomial distribution with
the given \code{expected.counts} (argument \code{mu}
to \code{\link[stats]{NegBinomial}}) and dispersion parameter
\code{nbinom.size}.}

\item{nbinom.size}{The dispersion parameter for the negative
binomial distribution; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}.}

\item{verbose}{If \code{TRUE} print messages under some circumstances.}
}
\value{
A \code{\link[tibble]{tibble}} with 
   self-explanatory columns and rows.
}
\description{
We use a separate function so as not to slow down the heavily
used \code{\link{LLHSpectrumNegBinom}} and to provide more
information in the output
}
