% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareBackgroundSignaturePackageData.R
\name{LoadArnoudMCF10HepG2}
\alias{LoadArnoudMCF10HepG2}
\title{Load spectra rceived from Arnoud on 2019 Oct 23}
\usage{
LoadArnoudMCF10HepG2()
}
\description{
Load spectra rceived from Arnoud on 2019 Oct 23
}
