% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{LogLHNegBinom}
\alias{LogLHNegBinom}
\title{Negative binomial log likelihood that an observed spectrum was generated from a vector of expected counts.}
\usage{
LogLHNegBinom(spectrum, expected.counts, nbinom.size)
}
\arguments{
\item{spectrum}{An observed spectrum (a numeric vector)}

\item{expected.counts}{A vector of expected mutation counts, one
expected count for each mutation type. We want to know the 
likelihood that this model generated the observed 
spectrum, assuming each mutational types generates counts according to
a negative binomial distribution with 
the given \code{expected.counts} (argument \code{m}
to \code{\link[stats]{dnbinom}}) and dispersion parameter
\code{nbinom.size}.}

\item{nbinom.size}{The \code{size} parameter that
governs dispersion. See \code{\link[stats]{dnbinom}}.
Smaller values correspond to larger dispersion.}
}
\value{
\code{log(likelihood(spectrum | expected.counts))}
}
\description{
Negative binomial log likelihood that an observed spectrum was generated from a vector of expected counts.
}
