% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareBackgroundSignaturePackageData.R
\name{MakeBackground}
\alias{MakeBackground}
\title{Estimate a background signature.}
\usage{
MakeBackground(bg.spectra, maxeval)
}
\description{
Estimate a background signature.
}
\keyword{internal}
