% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareBacgroundSignaturePackageData.R,
%   R/PreparePackageData.R
\name{MakeCisplatinCatalogs}
\alias{MakeCisplatinCatalogs}
\title{Make spectrum catalog from VCFs from cisplatin exposed HepG2}
\usage{
MakeCisplatinCatalogs()

MakeCisplatinCatalogs()
}
\value{
The catalog

The catalog
}
\description{
Make spectrum catalog from VCFs from cisplatin exposed HepG2

Make spectrum catalog from VCFs from cisplatin exposed HepG2
}
\keyword{internal}
