% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePackageData.R
\name{MakeHepG2BackgroundPart2}
\alias{MakeHepG2BackgroundPart2}
\title{Read the specified spectra file and estimate the HepG2 background signature.}
\usage{
MakeHepG2BackgroundPart2(maxeval)
}
\description{
Read the specified spectra file and estimate the HepG2 background signature.
}
\keyword{internal}
