% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeSynBackground}
\alias{MakeSynBackground}
\title{Generate num.samples background spectra}
\usage{
MakeSynBackground(background.info, num.samples)
}
\arguments{
\item{background.info}{}

\item{num.samples}{}
}
\description{
Generate num.samples background spectra
}
