% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeSynTestGrid}
\alias{MakeSynTestGrid}
\title{Make a "grid" of tests, with various "target signatures" and contribution from background}
\usage{
MakeSynTestGrid(sig.names.to.test, contribution, bg.sig.info,
  num.replicates = 10)
}
\value{
A \code{data.frame}, of which each row is a synthetic
spectrum, and the last 96 columns are the signatures.
}
\description{
Make a "grid" of tests, with various "target signatures" and contribution from background
}
\keyword{internal}
