% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeTestSpectrum}
\alias{MakeTestSpectrum}
\title{Make a single test spectrum.}
\usage{
MakeTestSpectrum(bg.sig.info, bg.contribution, target.sig = NULL)
}
\arguments{
\item{bg.sig.info}{See, for example, \code{HepG2.background.info}.}

\item{bg.contribution}{Expected proportion of mutations due to the background.}

\item{target.sig}{Target signature as a signature.}
}
\description{
Make a single test spectrum.
}
