% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateBackground.R
\name{MeanOfSpectraAsSig}
\alias{MeanOfSpectraAsSig}
\title{Return the mean of multiple spectra as a signature.}
\usage{
MeanOfSpectraAsSig(spectra)
}
\arguments{
\item{spectra}{Convert each spectrum to a signature and then compute the
mean of all signatures.}
}
\description{
Return the mean of multiple spectra as a signature.
}
