% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateBackground.R
\name{NegLLHOfSignature}
\alias{NegLLHOfSignature}
\title{Objective function for estimating a background signature from multiple background spectra.}
\usage{
NegLLHOfSignature(sig.and.nbinom.size, spectra)
}
\arguments{
\item{sig.and.nbinom.size}{Concatenation of signature as a vector and
the negative binomial dispersion parameter.For the dispersion parameter for the negative
       binomial distribution, smaller is more dispersed.
       See \code{\link[stats]{NegBinomial}}.}

\item{spectra}{The observed spectra as an \code{\link[ICAMS]{ICAMS}}
\code{catalog}.}
}
\value{
-1 * the log likelihood that the the input signature,
dispersion parameter, and mutation count generated 
the observed spectrum
}
\description{
Returns the negative log likelihood
that the signature with an
estimated negative binomial
dispersion parameter,
generated the observed (multiple) spectra given the
likelihood of the observed total counts
attributed to the signature.
}
\keyword{internal}
