% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrosamine.examples}
\alias{nitrosamine.examples}
\title{Example nitrosamine data for background subtraction}
\format{
A list of spectra catalogs for nitrosamines.
 The names of the catalogs are self-explanatory.
 Each catalog has 2 spectra from each of four different
 nitrosamines, "NDEA", "NDMA", "NPIP", "NPYR". The
 samples names also should be self-explanatory.
}
\usage{
nitrosamine.examples
}
\description{
Example nitrosamine data for background subtraction
}
\keyword{datasets}
