% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{Nloptr1Tumor}
\alias{Nloptr1Tumor}
\title{Optimize assignment of a fixed set of signature activities for a \strong{single} tumor.}
\usage{
Nloptr1Tumor(spectrum, sigs, m.opts = NULL, eval_f, ...)
}
\arguments{
\item{spectrum}{A single mutational spectrum a numeric vector or
single-column numeric matrix.}

\item{sigs}{A matrix of mutational signatures.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{eval_f}{See \code{\link[nloptr]{nloptr}}.}

\item{...}{Additional arguments for \code{eval_f}.}

\item{local.opts}{See \code{\link[nloptr]{nloptr}}.}
}
\description{
Optimize assignment of a fixed set of signature activities for a \strong{single} tumor.
}
\keyword{internal}
