% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{NumFromId}
\alias{NumFromId}
\title{Get the numerical parts of identifiers.}
\usage{
NumFromId(s)
}
\arguments{
\item{s}{A character vector.}
}
\value{
A vector, each element of which is the integer
corresponding to the first string of digits of an element of s.
}
\description{
Get the numerical parts of identifiers.
}
\details{
Not very sophisticated.
}
\examples{
\dontrun{
x<- c("SBS22", "SBS2", "SBS7b", "SBS7a")
NumFromId(x)
x[order(NumFromId(x))]
}
}
\keyword{internal}
