% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{obj.fun.nbinom.maxlh}
\alias{obj.fun.nbinom.maxlh}
\title{Objective function for \code{[SparseAssignActivity]}.}
\usage{
obj.fun.nbinom.maxlh(exp, spectrum, sigs, nbinom.size)
}
\description{
Objective function for \code{[SparseAssignActivity]}.
}
\keyword{internal}
