% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{ObjFn1}
\alias{ObjFn1}
\title{Estimate a signature minus a background signature}
\usage{
ObjFn1(est.target.sig.and.b, obs.spectra, bg.sig.info)
}
\description{
Let the input spectra be s1, s2, ...
}
\details{
Fig target.sig that maximize the likelihoods
max_(b1, target.sig){s1 | 
            b1, target.sig, background.sig, 
            background.sig.nbinom.size, background.sig.count.mu,
     background.sig.count.nbinom.size}
b1 * background.sig + (total-mut1 - b1) * target.sig * prob(b1), \cr
b2 * background.sig + (total-mut2 - b2) * target.sig * prob(b2), \cr
... \cr
\cr

Note: I guess you could estimate "background" and "target.sig"
at the same time, but then background might be slightly (?)
different for each set of input spectra. To do that would
include the likelhoods
background.s1 | background.sig, background.sig.nbinom.size
background.s2 | background.sig, background.sig.nbinom.size
and include a maximization over background.sig and
background.sig.nbinom.size
....
}
\keyword{internal}
