% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{ObjFnBinomMaxLHMustRound}
\alias{ObjFnBinomMaxLHMustRound}
\title{A deprecated negative binomial maximum likelihood objective function.}
\usage{
ObjFnBinomMaxLHMustRound(exp, spectrum, sigs, nbinom.size)
}
\arguments{
\item{exp}{The matrix of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}

\item{nbinom.size}{The dispersion parameter for the negative
binomial distribution; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}.}
}
\description{
Use \code{\link{ObjFnBinomMaxLHNoRoundOK}} instead.
}
\details{
This function will lead to errors in some situations
when the rounded reconstructed signature contains 0s for
mutations classes for which the target spectrum is > 0.
}
