% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFnBinomMaxLH2.R
\name{ObjFnBinomMaxLHNoRoundOK}
\alias{ObjFnBinomMaxLHNoRoundOK}
\title{A deprecated negative binomial maximum likelihood objective function.}
\usage{
ObjFnBinomMaxLHNoRoundOK(exp, spectrum, sigs, nbinom.size)
}
\arguments{
\item{exp}{A vector of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}

\item{nbinom.size}{The dispersion parameter for the negative
binomial distribution; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}.}
}
\description{
Use \code{\link{ObjFnBinomMaxLHRound}} instead.
}
\details{
This function rounds sometimes, which leads to
minor differences in log likelihoods of reconstructed spectra
(\code{\link{LLHSpectrumNegBinom}})
compared to the value returned by this function.
}
