% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFnBinomMaxLH2.R
\name{ObjFnBinomMaxLHRound}
\alias{ObjFnBinomMaxLHRound}
\title{The preferred negative binomial maximum likelihood objective function.}
\usage{
ObjFnBinomMaxLHRound(exp, spectrum, sigs, nbinom.size)
}
\arguments{
\item{exp}{A vector of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}

\item{nbinom.size}{The dispersion parameter for the negative
binomial distribution; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}.}
}
\value{
-1 * log(likelihood(spectrum | reconstruction))

\code{\link[nloptr]{nloptr}} minimizes the objective function, so the
lower the objective function, the better.
}
\description{
Can be used as the
objective function for \code{\link{MAPAssignActivity}},
\code{\link{MAPAssignActivity1}},
\code{\link{SparseAssignActivity}},
\code{\link{SignaturePresenceTest}}
and \code{\link{SignaturePresenceTest1}}.
(Internally used by by \code{\link[nloptr]{nloptr}}.)
}
\keyword{internal}
