% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFnMultinomMaxLH.R
\name{ObjFnMultinomMaxLH}
\alias{ObjFnMultinomMaxLH}
\title{The multinomial likelihood objective function}
\usage{
ObjFnMultinomMaxLH(exp, spectrum, sigs)
}
\arguments{
\item{exp}{The matrix of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}
}
\value{
-1 * log(likelihood(spectrum | reconstruction))

\code{\link[nloptr]{nloptr}} minimizes the objective function, so the
lower the objective function, the better.
}
\description{
Can be used as the
objective function for \code{\link{MAPAssignActivity}},
\code{\link{MAPAssignActivity1}},
\code{\link{SparseAssignActivity}},
\code{\link{SparseAssignActivity1}},
\code{\link{SignaturePresenceTest}}
and \code{\link{SignaturePresenceTest1}}.
(Internally used by by \code{\link[nloptr]{nloptr}}.)
}
\keyword{internal}
