% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{one.lh.and.exp}
\alias{one.lh.and.exp}
\title{Returns a list with elements:}
\usage{
one.lh.and.exp(spect, sigs, trace, algorithm = "NLOPT_LN_COBYLA",
  obj.fun, nbinom.size)
}
\description{
loglh    - the log likelihood of the best solution (set of exposures) found
exposure - the vector of exposures that generate loglh, in this case
exposure' means the number of mutations ascribed to each signature
}
\keyword{internal}
