% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMAPAssignTest.R
\name{OneMAPAssignTest}
\alias{OneMAPAssignTest}
\title{Run one test of MAPAssignActivity1}
\usage{
OneMAPAssignTest(
  spect,
  reference.exp,
  cancer.type,
  mutation.type,
  exposure.mutation.type,
  max.subsets = 1000,
  max.level = 5,
  max.mc.cores = 100,
  out.dir = NULL,
  p.thresh,
  max.presence.proportion
)
}
\arguments{
\item{spect}{xxxx.}

\item{reference.exp}{xxxxx}

\item{cancer.type}{xxxxx}

\item{mutation.type}{One of "SBS96", "SBS192", "ID", "DBS78".}

\item{exposure.mutation.type}{One of "SBS96", "ID", "DBS78".}

\item{max.subsets}{xxxxxx,}

\item{max.level}{XXXX}

\item{max.mc.cores}{XXXXX}

\item{out.dir}{XXXXXX}

\item{max.presence.proportion}{xxxxx}
}
\description{
Run one test of MAPAssignActivity1
}
