% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{P.of.M}
\alias{P.of.M}
\title{Calculate \eqn{P(M)} -- the probability of a model of which signatures are present in a sample.}
\usage{
P.of.M(model, sigs.presence.prop)
}
\arguments{
\item{model}{Names of sigs present in a trial exposure. Do not use indices.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a superset of \code{model}.}
}
\description{
Calculate \eqn{P(M)} -- the probability of a model of which signatures are present in a sample.
}
\keyword{internal}
