% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMAPAssignTest.R
\name{PCAWGMAPTest}
\alias{PCAWGMAPTest}
\title{Run MAPAssignSignature on one sample from PCAWG}
\usage{
PCAWGMAPTest(
  cancer.type,
  sample.index,
  mutation.type,
  max.level = 5,
  max.mc.cores
)
}
\arguments{
\item{cancer.type}{A cancer type from the exposures matrix.}

\item{sample.index}{The index of the sample within the exposures matrix.}

\item{mutation.type}{One of "SBS96" ...}
}
\description{
Run MAPAssignSignature on one sample from PCAWG
}
