% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherFunctions.R
\name{plot.recon.by.range}
\alias{plot.recon.by.range}
\title{Plot spectrum reconstruction broken up by ranges}
\usage{
\method{plot}{recon.by.range}(path, spect, sigs, ex, range, obj.fun,
  nbinom.size)
}
\description{
Plot spectrum reconstruction broken up by ranges
}
\keyword{internal}
