% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherFunctions.R
\name{plot.reconstruction}
\alias{plot.reconstruction}
\title{Plot information about reconstructed spectrum}
\usage{
\method{plot}{reconstruction}(signatures, exposures.mat, input.genomes, normalize.recon, padding = 0)
}
\description{
Plot information about reconstructed spectrum
}
\keyword{internal}
