% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{PlotExposureInternal}
\alias{PlotExposureInternal}
\title{Plot a single exposure plot}
\usage{
PlotExposureInternal(
  exposure,
  plot.proportion = FALSE,
  xlim = NULL,
  ylim = NULL,
  legend.x = NULL,
  legend.y = NULL,
  cex.legend = 0.9,
  cex.yaxis = 1,
  cex.xaxis = NULL,
  plot.sample.names = TRUE,
  yaxis.labels = NULL,
  ...
)
}
\arguments{
\item{exposure}{Exposures as a numerical \code{matrix} (or \code{data.frame})
with signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs. If you want
\code{exposure} sorted from largest to smallest, use
\code{\link{SortExposure}}. Do not use column names that start with
multiple underscores. The exposures will often be mutation counts, but
could also be e.g. mutations per megabase.}

\item{plot.proportion}{Plot exposure proportions rather than counts.}

\item{xlim, ylim}{Limits for the x and y axis. If \code{NULL}(default), the
function tries to do something reasonable.}

\item{legend.x, legend.y}{The x and y co-ordinates to be used to position the
legend.}

\item{cex.legend}{A numerical value giving the amount by which legend
plotting text and symbols should be magnified relative to the default.}

\item{cex.yaxis}{A numerical value giving the amount by which y axis values
should be magnified relative to the default.}

\item{cex.xaxis}{A numerical value giving the amount by which x axis values
should be magnified relative to the default. If
\code{NULL}(default), the function tries to do something reasonable.}

\item{plot.sample.names}{Whether to plot sample names below the x axis.
Default is TRUE.}

\item{yaxis.labels}{User defined y axis labels to be plotted. If
\code{NULL}(default), the function tries to do something reasonable.}

\item{...}{Other arguments passed to \code{\link[graphics]{barplot}}. If
\code{ylab} is not included, it defaults to a value depending on
\code{plot.proportion}. If \code{col} is not supplied the function tries to
do something reasonable.}
}
\value{
An \strong{invisible} list whose first element is a logic value
  indicating whether the plot is successful. The second element is a numeric
  vector giving the coordinates of all the bar midpoints drawn, useful for
  adding to the graph.
}
\description{
Plot a single exposure plot
}
\keyword{internal}
