% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowSigActivity.R
\name{PlotSigActivityToPdf}
\alias{PlotSigActivityToPdf}
\title{Plot signature activity from the output generated by \code{AddSigActivity} to PDF}
\usage{
PlotSigActivityToPdf(
  list.of.sig.activity,
  output.dir,
  plot.all.samples.in.one.pdf = TRUE,
  ...
)
}
\arguments{
\item{list.of.sig.activity}{A list of contributing signature activity
information for multiple spectra. See the return value of
\code{\link{AddSigActivity}} for more details.}

\item{output.dir}{The directory to save the results. Create
this directory if it does not exist.}

\item{plot.all.samples.in.one.pdf}{Whether to plot all the signature activity information
within one PDF. Default is TRUE. If FALSE, then plot one PDF for each sample.}

\item{...}{Other arguments passed to \code{\link[ICAMS]{PlotCatalogToPdf}}.}
}
\value{
An \strong{invisible} list whose first element is a logic value
  indicating whether the plot is successful.
}
\description{
Plot signature activity from the output generated by \code{AddSigActivity} to PDF
}
\keyword{internal}
