% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{Polish}
\alias{Polish}
\title{"Polish" a solution by minimizing Euclidean distance to the input spectrum.}
\usage{
Polish(exp, sig.names, spect)
}
\description{
This is experimental for testing.
}
\keyword{internal}
