% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureProportions.R
\name{PossibleArtifacts}
\alias{PossibleArtifacts}
\title{Return a character vector of the IDs of possible SBS96 signature artifacts.}
\usage{
PossibleArtifacts()
}
\description{
Return a character vector of the IDs of possible SBS96 signature artifacts.
}
