% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAssignActivity.R
\name{PresenceAssignActivity}
\alias{PresenceAssignActivity}
\title{Find minimal set of signatures that can explain multiple spectra by first using
signature presence test}
\usage{
PresenceAssignActivity(
  spectra,
  sigs,
  output.dir,
  p.thresh = DefaultPThresh(sigs),
  m.opts = DefaultManyOpts(spectra = spectra),
  num.parallel.samples = 1,
  mc.cores.per.sample = 1,
  seed = 123,
  drop.low.mut.samples = FALSE,
  save.files = TRUE
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{output.dir}{Directory path to save the output file.}

\item{p.thresh}{If
the p value for a better reconstruction with a
set of signatures (as opposed to
without that set of signatures)
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{num.parallel.samples}{The (maximum) number of samples to run in
parallel. On Microsoft Windows machines it is silently changed to 1. Each
sample in turn can require multiple cores, as governed by
\code{mc.cores.per.sample}.}

\item{mc.cores.per.sample}{The maximum number of cores to use for each
sample. On Microsoft Windows machines it is silently changed to 1.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}

\item{drop.low.mut.samples}{Whether to exclude low mutation samples from
the analysis. If \code{TRUE}, samples with SBS total mutations less
than 100, DBS or ID total mutations less than 25 will be dropped.}

\item{save.files}{If \code{TRUE} save several files for each input sample
in a directory named after the sample.}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: The proposed set of signatures that can
plausibly explain \code{spectra}.
\item \code{proposed.reconstruction}: The reconstruction based on
\code{proposed.assignment}.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spectra} and \code{proposed.reconstruction}.
\item \code{time.for.assignment}: Value from \code{system.time} for running
\code{PresenceAssignActivity} for each sample in \code{spectra}.
\item \code{error.messages}: Error messages running
\code{PresenceAssignActivity}.
}
}
\description{
Find minimal set of signatures that can explain multiple spectra by first using
signature presence test
}
\examples{
\dontrun{
# This is a long running example unless parallel computing is supported on your machine
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
SBS96.sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
sigs.to.use <- SBS96.sigs[, names(sigs.prop), drop = FALSE]
retval <- PresenceAssignActivity(spectra = spectra,
                                 sigs = sigs.to.use,
                                 output.dir = file.path(tempdir(), "Lung-AdenoCA"),
                                 num.parallel.samples = 2,
                                 mc.cores.per.sample = 30,
                                 seed = 2561)
}
}
