% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{prop.reconstruct}
\alias{prop.reconstruct}
\title{error checked (?) function to get reconstructed something?}
\usage{
prop.reconstruct(sigs, exp)
}
\description{
error checked (?) function to get reconstructed something?
}
\keyword{internal}
