% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureProportions.R
\name{RareSignatures}
\alias{RareSignatures}
\title{Return a character vector of the IDs of rare SBS96 signatures.}
\usage{
RareSignatures()
}
\description{
Return a character vector of the IDs of rare SBS96 signatures.
}
