% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{ReadExposure}
\alias{ReadExposure}
\title{Read an exposure matrix from a file}
\usage{
ReadExposure(file, check.names = FALSE)
}
\arguments{
\item{file}{CSV file containing an exposure matrix.}

\item{check.names}{Passed to \code{read.csv}.
\strong{IMPORTANT}: If \code{TRUE} this will replace the double
colon in identifiers of the form <tumor_type>::<sample_id>
with two periods (i.e. <tumor_type>..<sample_id>.
If \code{check.names} is true, generate a warning
if double colons were present.}
}
\value{
Matrix of exposures.
}
\description{
Read an exposure matrix from a file
}
\examples{
file <- system.file("extdata",
                    "synthetic.exposure.csv",
                    package = "mSigAct")
exposure <- ReadExposure(file)
}
