% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMAPOneSample.R
\name{RunMAPOnOneSample}
\alias{RunMAPOnOneSample}
\title{Run \code{MAPAssignActivity1} on one sample and save and plot the results}
\usage{
RunMAPOnOneSample(
  spect,
  sigs,
  sigs.presence.prop,
  output.dir,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  max.mc.cores = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL,
  max.subsets = 1000,
  use.sparse.assign = FALSE,
  drop.low.mut.samples = TRUE,
  use.sig.presence.test = FALSE,
  sig.pres.test.nbinom.size = NULL,
  sig.pres.test.p.thresh = 0.05,
  sig.pres.test.q.thresh = NULL,
  save.files
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{output.dir}{Directory path to save the output file.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with a
set of signatures (as opposed to
without that set of signatures)
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}

\item{max.subsets}{This argument provides a way to 
heuristically limit the
amount of time spent by this function. Larger values of this
argument will tend to allow longer running times.
The algorithm
successively tries to remove all subsets of 1 signature, 2
signatures, 3 signatures, etc., down to \code{max.level}.
(Not every subset is tested at each level; if a subset was
already found to be necessary the algorithm does not test
supersets of that subset.) If at any level the algorithm
needs to test more than \code{max.subsets} this function will
not proceed.}

\item{use.sparse.assign}{Whether to use sparse assignment. If \code{TRUE},
arguments designed for Maximum A Posteriori assignment such as
\code{sigs.presence.prop} will be ignored.}

\item{drop.low.mut.samples}{Whether to exclude low mutation samples from
the analysis. If \code{TRUE(default)}, samples with SBS total mutations less
than 100, DBS or ID total mutations less than 25 will be dropped.}

\item{use.sig.presence.test}{Whether to use signature presence test first to
filter out those signatures that are not needed in the reconstruction of
the spectrum.}

\item{sig.pres.test.nbinom.size}{Only needed when \code{use.sig.presence.test
= TRUE}. The dispersion parameter for the negative binomial distribution
used when conducting signature presence test; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}. If \code{NULL}, then use multinomial
likelihood when conducting signature presence test.}

\item{sig.pres.test.p.thresh}{Only needed when \code{use.sig.presence.test
= TRUE}. If the p-value of signature presence test for one signature is
>= \code{sig.pres.test.p.thresh}, then this signature will not be used for
assignment later.}
}
\description{
Run \code{MAPAssignActivity1} on one sample and save and plot the results
}
\keyword{internal}
