% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{RunTests}
\alias{RunTests}
\title{Run a suite of tests given data specfied in \code{test.table}}
\usage{
RunTests(test.table, num.replicates, num.spectra.per.replicate, bg.info,
  out.dir, mc.cores = 10, maxeval = 1000, xtol_rel = 0.001/10,
  xtol_abs = 1e-04/10, print_level = 0,
  algorithm = "NLOPT_LN_COBYLA", verbose = FALSE)
}
\arguments{
\item{test.table}{XXX}

\item{num.replicates}{Currently ignored and assumed to be 1.}

\item{num.spectra.per.replicate}{The number of spectra from which to subtract
the background signature. For the Duke-NUS 
workflow, 2 or 3 would be a realistic number.
Currently we just take the first 2 spectra
for a given target signature and background.
In the future we could take multiple samples.}

\item{bg.info}{Information on the background signatures, e.g. \code{HepG2.background.info}.}

\item{out.dir}{XXXX}

\item{mc.cores}{See \code{FindSignatureMinusBackground}.}

\item{maxeval}{See \code{FindSignatureMinusBackground}.}

\item{xtol_rel}{See \code{FindSignatureMinusBackground}.}

\item{xtol_abs}{See \code{FindSignatureMinusBackground}.}

\item{print_level}{See \code{FindSignatureMinusBackground}.}

\item{algorithm}{See \code{FindSignatureMinusBackground}.}

\item{verbose}{If \code{TRUE} print some progress information.}
}
\description{
Run a suite of tests given data specfied in \code{test.table}
}
