% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowSigActivity.R
\name{ShowExposure}
\alias{ShowExposure}
\title{Show signature exposure from the output generated by \code{AddSigActivity}}
\usage{
ShowExposure(
  list.of.sig.activity,
  output.dir,
  base.filename,
  mutation.type,
  plot.exposure.proportion = FALSE,
  mfrow = c(2, 1),
  mar = c(6, 4, 3, 2),
  oma = c(3, 2, 0, 2)
)
}
\arguments{
\item{list.of.sig.activity}{A list of contributing signature activity
information for multiple spectra. See the return value of
\code{\link{AddSigActivity}} for more details.}

\item{output.dir}{The directory to save the results. Create
this directory if it does not exist.}

\item{base.filename}{Optional. \code{base.filename} will be appended to the start
of the names of files generated inside \code{output.dir}.}

\item{mutation.type}{The mutation type of the mutational spectrum.}

\item{plot.exposure.proportion}{Whether to plot exposure proportions rather
than counts.}

\item{mfrow}{A vector of the form \code{c(nr, nc)}. Subsequent figures will
be drawn in an \code{nr}-by-\code{nc} array on the device by rows.}

\item{mar}{A numerical vector of the form \code{c(bottom, left, top, right)}
which gives the number of lines of margin to be specified on the four sides
of the plot.}

\item{oma}{A vector of the form \code{c(bottom, left, top, right)} giving the
size of the outer margins in lines of text.}
}
\description{
Show signature exposure from the output generated by \code{AddSigActivity}
}
\keyword{internal}
