% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowSigActivity.R
\name{ShowSigActivity}
\alias{ShowSigActivity}
\title{Show signature activity from the output generated by \code{AddSigActivity}}
\usage{
ShowSigActivity(
  list.of.sig.activity,
  output.dir,
  base.filename = NULL,
  plot.all.samples.in.one.pdf = TRUE,
  plot.exposure.proportion = FALSE,
  ...
)
}
\arguments{
\item{list.of.sig.activity}{A list of contributing signature activity
information for multiple spectra. See the return value of
\code{\link{AddSigActivity}} for more details.}

\item{output.dir}{The directory to save the results. Create
this directory if it does not exist.}

\item{base.filename}{Optional. \code{base.filename} will be appended to the start
of the names of files generated inside \code{output.dir}.}

\item{plot.all.samples.in.one.pdf}{Whether to plot all the signature activity information
within one PDF. Default is TRUE. If FALSE, then plot one PDF for each sample.}

\item{plot.exposure.proportion}{Whether to plot exposure proportions rather
than counts.}

\item{...}{Other arguments passed to \code{\link[ICAMS]{PlotCatalogToPdf}}.}
}
\description{
Show signature activity from the output generated by \code{AddSigActivity}
}
\examples{
\dontrun{
spectra <- PCAWG7::spectra$PCAWG$SBS96[, 1:2, drop = FALSE]
exposure <- PCAWG7::exposure$PCAWG$SBS96[, 1:2, drop = FALSE]
sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37
sigs.prop <- ExposureProportions(mutation.type = "SBS96", 
                                 cancer.type = "Biliary-AdenoCA")
retval <- AddSigActivity(spectra, exposure, sigs, sigs.prop)
ShowSigActivity(retval, output.dir = file.path(tempdir(), "SBS96"), 
                base.filename = "Biliary-AdenoCA")
}                
}
\keyword{internal}
