% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturePresenceTest.R
\name{SignaturePresenceTest}
\alias{SignaturePresenceTest}
\title{Test whether a given signature is plausibly present in a catalog.}
\usage{
SignaturePresenceTest(
  spectra,
  sigs,
  target.sig.index,
  m.opts = NULL,
  mc.cores = 10
)
}
\arguments{
\item{spectra}{The catalog (matrix) to analyze. This could be
an \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{sigs}{A catalog of signatures from which to choose.
This could be
  and \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{target.sig.index}{The index of the signature the presence
of which we want to test.}

\item{m.opts}{If \code{NULL} use the return from
calling \code{\link{DefaultManyOpts}}. For documentation
see \code{\link{DefaultManyOpts}}.}

\item{mc.cores}{Number of cores to use. Always silently
changed to 1 on Microsoft Windows.}
}
\description{
Test whether a given signature is plausibly present in a catalog.
}
\examples{
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
sigs <- PCAWG7::signature$genome$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
sigs.to.use <- sigs[, names(sigs.prop), drop = FALSE]
# Test whether SBS17a is plausibly present in the spectra
sig.index <- which(colnames(sigs.to.use) == "SBS17a")
sig.presence.test.out <- SignaturePresenceTest(spectra = spectra,
                                               sigs = sigs.to.use, 
                                               target.sig.index = sig.index,
                                               mc.cores = 2)
}
