% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturePresenceTest1.R
\name{SignaturePresenceTest1}
\alias{SignaturePresenceTest1}
\title{Test whether a given signature is plausibly present in a spectrum.}
\usage{
SignaturePresenceTest1(
  spectrum,
  sigs,
  target.sig.index,
  m.opts = DefaultManyOpts(),
  seed = NULL
)
}
\arguments{
\item{spectrum}{The spectrum to analyze.}

\item{sigs}{A catalog of signatures from which to choose.}

\item{target.sig.index}{The index of the signature the presence
of which we want to test. It can also be the signature id (e.g. "SBS22").}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}
}
\value{
A list of elements:
\itemize{
\item loglh.with: The maximum log likelihood of the reconstructed spectrum using
all the signatures.
\item loglh.without: The maximum log likelihood of the reconstructed spectrum
without the target signature.
\item statistic: Likelihood ratio test statistic.
\item chisq.p: P-value of the likelihood ratio test. The null hypothesis is we
can plausibly reconstruct the spectrum without the target signature.
\item exp.with: The exposure using all the signatures which generates the maximum
log likelihood \code{loglh.with}.
\item exp.without: The exposure not using the target signature which generates
the maximum log likelihood \code{loglh.without}.
}
}
\description{
For backward compatibility. See also \code{\link{AnySigSubsetPresent}}.
}
\keyword{internal}
