% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPresenceAssignActivity.R
\name{SigPresenceAssignActivity}
\alias{SigPresenceAssignActivity}
\title{Find minimal set of signatures that can explain multiple spectra by first using
signature presence test}
\usage{
SigPresenceAssignActivity(
  spectra,
  sigs,
  output.dir,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  num.parallel.samples = 5,
  mc.cores.per.sample = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL,
  drop.low.mut.samples = TRUE,
  sig.pres.test.nbinom.size = NULL,
  sig.pres.test.p.thresh = 0.05,
  sig.pres.test.q.thresh = NULL,
  save.files = TRUE
)
}
\arguments{
\item{sig.pres.test.nbinom.size}{The dispersion parameter for the negative
binomial distribution used when conducting signature presence test first to
filter out those signatures that are not needed in the reconstruction of
the spectrum; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}. If \code{NULL}, then use multinomial
likelihood when conducting signature presence test.}

\item{sig.pres.test.p.thresh}{If the p-value of signature presence test for
one signature is >= \code{sig.pres.test.p.thresh}, then this signature will
not be used for assignment later.}

\item{sig.pres.test.q.thresh}{A test parameter}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: The proposed set of signatures that can
plausibly explain \code{spectra}.
\item \code{proposed.reconstruction}: The reconstruction based on
\code{proposed.assignment}.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spectra} and \code{proposed.reconstruction}.
\item \code{all.tested}: All tested possible ways to reconstruct each
sample in \code{spectra}.
\item \code{alt.solutions}: A \code{tibble} showing all the alternative solutions
that are statistically as good as the \code{proposed.assignment} that can
plausibly reconstruct \code{spectra}.
\item \code{time.for.assignment}: Value from \code{system.time} for running
\code{SigPresenceAssignActivity} for each sample in \code{spectra}.
\item \code{error.messages}: Only appearing if there are errors running
\code{SigPresenceAssignActivity}.
}

The elements \code{proposed.assignment}, \code{proposed.reconstruction},
\code{reconstruction.distances}, \code{all.tested},
\code{time.for.assignment} will be \code{NULL} if the algorithm could not
find the optimal reconstruction or there are errors coming out for
\strong{all} samples.
}
\description{
Find minimal set of signatures that can explain multiple spectra by first using
signature presence test
}
\examples{
\dontrun{
# This is a long running example unless parallel computing is supported on your machine
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
SBS96.sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
sigs.to.use <- SBS96.sigs[, names(sigs.prop), drop = FALSE]
retval <- SigPresenceAssignActivity(spectra = spectra,
                                    sigs = sigs.to.use,
                                    output.dir = file.path(tempdir(), "Lung-AdenoCA"),
                                    max.level = ncol(sigs.to.use) - 1,
                                    p.thresh = 0.05 / ncol(sigs.to.use),
                                    num.parallel.samples = 2,
                                    mc.cores.per.sample = 30,
                                    seed = 2561)
}
}
