% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple.100000.NLOPT_LN_COBYLA}
\alias{simple.100000.NLOPT_LN_COBYLA}
\title{Another temp test output}
\format{An object of class \code{list} of length 12.}
\usage{
simple.100000.NLOPT_LN_COBYLA
}
\description{
Another temp test output
}
\keyword{datasets}
