% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple.200000.NLOPT_GN_DIRECT_L}
\alias{simple.200000.NLOPT_GN_DIRECT_L}
\title{Another temp test ouput}
\format{An object of class \code{list} of length 12.}
\usage{
simple.200000.NLOPT_GN_DIRECT_L
}
\description{
Another temp test ouput
}
\keyword{datasets}
