% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple.40000.HepG2.tests}
\alias{simple.40000.HepG2.tests}
\title{Temporary test data}
\format{Complicated}
\usage{
simple.40000.HepG2.tests
}
\description{
Temporary test data
}
\examples{
# XXXXXX

}
\keyword{datasets}
