% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple.40000.remainder}
\alias{simple.40000.remainder}
\title{Temporary test data}
\format{Complicated}
\usage{
simple.40000.remainder
}
\description{
Temporary test data
}
\examples{
# XXXXXX
}
\keyword{datasets}
