% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{SortExposure}
\alias{SortExposure}
\title{Sort columns of an exposure matrix from largest to smallest (or vice versa)}
\usage{
SortExposure(exposure, decreasing = TRUE)
}
\arguments{
\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs.}

\item{decreasing}{If \code{TRUE}, sort from largest to smallest.}
}
\value{
The original \code{exposure} with columns sorted.
}
\description{
Sort columns of an exposure matrix from largest to smallest (or vice versa)
}
\examples{
file <- system.file("extdata",
                    "synthetic.exposure.csv",
                    package = "ICAMSxtra")
exposure <- ReadExposure(file)
exposure.sorted <- SortExposure(exposure)
}
