% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp.sigs.exome}
\alias{sp.sigs.exome}
\title{SigProfiler 96 SBS \code{counts.signatures} under exome abundance.}
\format{An \code{[ICAMS]} \code{counts.spectrum} catalog;
 see Alexandrov et al. 
 \url{https://www.biorxiv.org/content/10.1101/322859v2}.}
\usage{
sp.sigs.exome
}
\description{
SigProfiler 96 SBS \code{counts.signatures} under exome abundance.
}
\keyword{datasets}
