% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp.sigs}
\alias{sp.sigs}
\title{SigProfiler 96 SBS signatures}
\format{An \code{[ICAMS]} \code{counts.spectrum} catalog; see Alexandrov et al. ?????.}
\usage{
sp.sigs
}
\description{
SigProfiler 96 SBS signatures
}
\keyword{datasets}
