% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignActivity1OLD}
\alias{SparseAssignActivity1OLD}
\title{Component of \code{\link{SparseAssignActivity}} for one spectrum.}
\usage{
SparseAssignActivity1OLD(
  spect,
  sigs,
  max.level = 5,
  p.thresh = 0.05,
  eval_f,
  m.opts
)
}
\description{
Component of \code{\link{SparseAssignActivity}} for one spectrum.
}
\keyword{internal}
