% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignTest}
\alias{SparseAssignTest}
\title{Framework for testing \code{\link{SparseAssignActivity}}.}
\usage{
SparseAssignTest(sig.counts, trace = 0, mc.cores = NULL)
}
\arguments{
\item{sig.counts}{A matrix of target exposures:
rows are signatures and columns are tumors.}

\item{trace}{If > 0 print progress messages.}
}
\description{
Framework for testing \code{\link{SparseAssignActivity}}.
}
\keyword{internal}
