% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignTest1}
\alias{SparseAssignTest1}
\title{Framework for testing \code{\link{SparseAssignActivity1}}.}
\usage{
SparseAssignTest1(sig.counts, trace = 0)
}
\description{
Tests only for single spectrum.
}
\keyword{internal}
