% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignTest1OLD}
\alias{SparseAssignTest1OLD}
\title{Framework for testing \code{\link{SparseAssignActivity1}}.}
\usage{
SparseAssignTest1OLD(sig.counts, trace = 0, max.mc.cores = 1)
}
\description{
Tests only for single spectrum.
}
\keyword{internal}
