% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignTestGeneric}
\alias{SparseAssignTestGeneric}
\title{Framework for testing \code{\link{SparseAssignActivity1}}.}
\usage{
SparseAssignTestGeneric(sig.counts, trace = 0)
}
\description{
Framework for testing \code{\link{SparseAssignActivity1}}.
}
\keyword{internal}
