% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsForSubtractingBackground.R
\name{Test0}
\alias{Test0}
\title{Test \code{FindSignatureMinusBackground} on background-only spectra.}
\usage{
Test0(start.b.fraction = 0.9, maxeval = 10000)
}
\description{
Test \code{FindSignatureMinusBackground} on background-only spectra.
}
\keyword{internal}
