% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{TestAltSolutions}
\alias{TestAltSolutions}
\title{Test to try to find alternative solutions that are statistically as good as
the best solution from \code{MAPAssignActivitiy1}}
\usage{
TestAltSolutions(tibble, sparse.assign = FALSE)
}
\description{
Test to try to find alternative solutions that are statistically as good as
the best solution from \code{MAPAssignActivitiy1}
}
\keyword{internal}
