% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturePresenceTest1.R
\name{TestSignaturePresenceTest1}
\alias{TestSignaturePresenceTest1}
\title{Framework for testing \code{\link{SignaturePresenceTest1}}.}
\usage{
TestSignaturePresenceTest1(
  sig.counts,
  input.sigs = cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96,
  trace = 0,
  m.opts = NULL
)
}
\description{
Framework for testing \code{\link{SignaturePresenceTest1}}.
}
\keyword{internal}
