% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{ThisObjFn}
\alias{ThisObjFn}
\title{Estimate a signature minus a background signature}
\usage{
ThisObjFn(est.target.sig.and.total.mut.and.b, target.spectra,
  background.sig.info)
}
\description{
Let the input spectra be s1, s2, ...
}
\keyword{internal}
