% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSparseAssignOneSample.R
\name{tryingtoremove}
\alias{tryingtoremove}
\title{Run \code{SparseAssignActivity1} on one sample and save and plot the results}
\arguments{
\item{output.dir}{Directory path to save the output file.}
}
\description{
Run \code{SparseAssignActivity1} on one sample and save and plot the results
}
\keyword{internal}
