% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSigActivity.R
\name{AddSigActivity1}
\alias{AddSigActivity1}
\title{Add contributing signature activity information for one sample}
\usage{
AddSigActivity1(spect, exposure, sigs, nbinom.size = 5)
}
\arguments{
\item{spect}{A single spectrum.}

\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs.}

\item{sigs}{The signatures with which we are trying to reconstruct \code{spect}.
A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog. The column
names of \code{sigs} should be a superset of row names of \code{exposure}.}

\item{nbinom.size}{The dispersion parameter for the negative binomial
distribution; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}.}
}
\value{
A list of elements:
\itemize{
\item \code{original.spect}: The original \code{spect} with total mutation counts
added to its column name. An additional attribute "exposure" from
\code{exposure} is also added.
\item \code{reconstructed.spect}: The reconstructed spectrum using \code{sigs}
and \code{exposure}. Its column name has the total mutation counts and cosine
similarity with the original \code{spect}.
\item \code{contributing.sigs}: The contributing signatures to the original
\code{spect}. The column names of each contributing signature has mutation
counts attributed to this signature, its contribution proportion and proposed
etiology(if the etiology is unknown, then will be blank.)
\item \code{distances}: Various distances and similarities between \code{spect}
and \code{\code{reconstructed.spect}}.
}
}
\description{
Add contributing signature activity information for one sample
}
\details{
This function calls \code{\link{ReconstructSpectrum}} and
\code{\link{LLHSpectrumNegBinom}}.
}
\section{Remark}{
 The column names of \code{spect} should be the same as the
column name of \code{exposure}.
}

\keyword{internal}
