% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{AnySigSubsetPresent}
\alias{AnySigSubsetPresent}
\title{Deterimine whether any of several signatures in any combination are plausibly needed to reconsruct a given spectrum.}
\usage{
AnySigSubsetPresent(
  spect,
  all.sigs,
  target.sigs.index,
  eval_f = mSigAct::ObjFnBinomMaxLHNoRoundOK,
  m.opts
)
}
\arguments{
\item{spect}{The spectrum to be reconstructed, as single column matrix or
\code{\link[ICAMS]{ICAMS}} catalog.}

\item{all.sigs}{The matrix or catalog of singantures of possible interest;
\code{all.sigs} includes the signatures that might or might not be necessary
for plausibly reconstructing \code{spect}; \code{all.sigs} includes the
signatures in \eqn{H_0} plus the additional signatures in \eqn{H_a}.}

\item{target.sigs.index}{An integer vector of the indices of the signatures
of interest (those in \eqn{H_a - H_0}.}

\item{eval_f}{XXXXX}

\item{m.opts}{XXXX

Return the P value of the null hypothesis that none of the signatures in Ha.sigs,}
}
\description{
Deterimine whether any of several signatures in any combination are plausibly needed to reconsruct a given spectrum.
}
