% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.info}
\alias{background.info}
\alias{HepG2.background.info}
\title{Specifications of background signatures}
\format{A list with the elements MORE MORE MORE}
\usage{
HepG2.background.info
}
\description{
Specifications of background signatures
}
\examples{
# MORE MORE MORE

}
\keyword{datasets}
