% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.info}
\alias{background.info}
\alias{HepG2.background.info}
\alias{MCF10A.background.info}
\title{Specifications of background signatures}
\format{
A list with the elements \describe{
\item{\code{background.sig}}{The background signature profile.}

\item{code{sig.nbinom.size}}{The dispersion parameter for the negative
       binomial distribution for sampling error around
       the components of \code{background.sig}. Smaller is more dispersed.
       See \code{\link[stats]{NegBinomial}}.}

\item{code{count.nbinom.mu}}{The \code{mu} argument
 for
\code{\link[stats]{NegBinomial}} for the distribution
of total counts due to \code{background.sig} across
replicate exposed clones.}

\item{\code{count.nbinom.size}}{The dispersion parameter
 for the negative binomial distribution of sampling error
 for total counts due to \code{background.sig} across
 replicate exposed clones. Smaller is more dispersed.
 See \code{\link[stats]{NegBinomial}}}

}

An object of class \code{list} of length 5.
}
\source{
\code{HepG2.background.info} was estimated from \code{\link{HepG2.background.spectra}}.

\code{MCF10A.background.info} was estimated from \code{\link{MCF10A.background.spectra}}
}
\usage{
HepG2.background.info

MCF10A.background.info
}
\description{
Specifications of background signatures

Background information for MCF-10A cells.
}
\examples{
HepG2.background.info$count.nbinom.mu
HepG2.background.info$count.nbinom.size
HepG2.background.info$sig.nbinom.size
HepG2.background.info$background.sig[1:3, ]



}
\keyword{datasets}
