% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.info}
\alias{background.info}
\alias{HepG2.background.info}
\alias{MCF10A.background.info}
\alias{kucab.background.info}
\title{Specifications of background signatures}
\format{A list with the elements \enumerate{
\item \code{background.sig} The background signature profile.

\item \code{sig.nbinom.size} The \code{size} argument for
\code{\link[stats]{NegBinomial}} for sampling error around
the components of \code{background.sig}.

\item \code{count.nbinom.mu} The \code{mu} argument
 for
\code{\link[stats]{NegBinomial}} for the distribution
of total counts due to \code{background.sig} across
replicate exposed clones.

\item \code{count.nbinom.size} The \code{size} argument
 for
\code{\link[stats]{NegBinomial}} for the distribution
of total counts due to \code{background.sig} across
replicate exposed clones.

}}
\source{
\code{HepG2.background.info} was estimated from \code{\link{HepG2.background.spectra}}.

\code{MCF10A.background.info} was estimated from \code{\link{MCF10A.background.spectra}}

\code{kucab.background.info} was estimated from the control spectra in \code{\link{kucab.control.spectra.ICAMS}}.
}
\usage{
HepG2.background.info

MCF10A.background.info

kucab.background.info
}
\description{
Specifications of background signatures

Background information for MCF-10A cells.

Background.info for cells used in Kucab et al, 2019.
}
\examples{
HepG2.background.info$count.nbinom.mu
HepG2.background.info$count.nbinom.size
HepG2.background.info$sig.nbinom.size
HepG2.background.info$background.sig[1:3, ]



}
\keyword{datasets}
