% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.info}
\alias{background.info}
\alias{HepG2.background.info}
\title{Specifications of background signatures}
\format{A list with the elements \enumerate{
\item \code{background.sig} The background signature profile.

\item \code{mean.sig} Deprecated -- the background signature
profile computed as the mean of the background spectra; this
is essentially the same as \code{background.sig}.

\item \code{sig.nbinom.size} The \code{size} argument for
\code{\link[stats]{dnbinom}} for sampling error around
the components of \code{background.sig}.

\item \code{count.nbinom.mu} The \code{mu} argument
 for
\code{\link[stats]{dnbinom}} for the distribution
of total counts due to \code{background.sig} across
replicate exposed clones.

\item \code{count.nbinom.size} The \code{size} argument
 for
\code{\link[stats]{dnbinom}} for the distribution
of total counts due to \code{background.sig} across
replicate exposed clones.

}}
\usage{
HepG2.background.info
}
\description{
Specifications of background signatures
}
\examples{
# MORE MORE MORE

}
\keyword{datasets}
