% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background.spectra}
\alias{background.spectra}
\alias{HepG2.background.spectra}
\alias{MCF10A.background.spectra}
\title{Background spectra for HepG2 and MCF-10A}
\format{An \code{\link[ICAMS]{ICAMS}} \code{counts} catalog.}
\usage{
HepG2.background.spectra

MCF10A.background.spectra
}
\description{
Background spectra for HepG2 and MCF-10A

Background spectra for MCF-10A cells
}
\keyword{datasets}
