% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisplatin.exposed.HepG2.96}
\alias{cisplatin.exposed.HepG2.96}
\title{Mutational spectra of cisplatin exposed HepG2 cells.}
\format{An \code{\link[ICAMS]{ICAMS}} \code{counts} catalog.}
\usage{
cisplatin.exposed.HepG2.96
}
\description{
Mutational spectra of cisplatin exposed HepG2 cells.
}
\keyword{datasets}
