% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultManyOpts.R
\name{DefaultManyOpts}
\alias{DefaultManyOpts}
\title{Set default options for many functions, especially \code{\link[nloptr]{nloptr}}}
\usage{
DefaultManyOpts(likelihood.dist = "multinom")
}
\arguments{
\item{likelihood.dist}{The probability distribution used to calculate the
likelihood, can be either "multinom" (multinomial distribution),
"neg.binom" (negative binomial distribution) or 
"dirichlet.multinom" (Dirichlet-multinomial distribution).}
}
\value{

}
\description{
Set default options for many functions, especially \code{\link[nloptr]{nloptr}}
}
\examples{
my.opts <- DefaultManyOpts()
my.opts$trace <- 10
}
