% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{DefaultManyOpts}
\alias{DefaultManyOpts}
\title{Set default options for many functions, especially \code{\link[nloptr]{nloptr}}.}
\usage{
DefaultManyOpts()
}
\value{
A list with the following elements
\describe{
  \item{global.opts}{Options for \code{\link[nloptr]{nloptr}},
  q.v., for the global optimization phase.}

  \item{local.opts}{Options for \code{\link[nloptr]{nloptr}},
  q.v., for the local optimization phase.}

  \item{nbinom.size}{The dispersion parameter for the negative
       binomial distribution; smaller is more dispersed.
       See \code{\link[stats]{NegBinomial}}.}

  \item{trace}{If > 0 print progress messages.}
}
}
\description{
Set default options for many functions, especially \code{\link[nloptr]{nloptr}}.
}
