% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureProportions.R
\name{ExposureProportions}
\alias{ExposureProportions}
\title{Return the proportions of tumors of a given cancer type that have a particular signature}
\usage{
ExposureProportions(
  mutation.type,
  cancer.type,
  all.sigs = NULL,
  drop.sigs.no.info = TRUE
)
}
\arguments{
\item{mutation.type}{A character string, one of "SBS96", "SBS192", "ID", "DBS78".}

\item{cancer.type}{A character string.}

\item{all.sigs}{An optional matrix of known signatures, with column names being signatures ids.}

\item{drop.sigs.no.info}{If TRUE, drop any not present in the column names of \code{all.sigs}.
There are some signatures that do not have SBS192 versions, including SBS29.}
}
\value{
A numerical vector of the proportion of tumors of type \code{cancer.type} with each signature
  for those signatures observed in \code{cancer.type}.The names are the signature ids.
}
\description{
Return the proportions of tumors of a given cancer type that have a particular signature
}
