% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubtractBackground.R
\name{FindSignatureMinusBackground}
\alias{FindSignatureMinusBackground}
\title{Estimate a signature from experimentally exposed spectra minus a background signature.}
\usage{
FindSignatureMinusBackground(
  spectra,
  bg.sig.info,
  m.opts = NULL,
  start.b.fraction = 0.1
)
}
\arguments{
\item{spectra}{The spectra from which to subtract the background,
as a matrix or \code{\link[ICAMS]{ICAMS}} catalog.}

\item{bg.sig.info}{Information about the background signature. See for example
\code{\link{HepG2.background.info}}.}

\item{m.opts}{Options to pass to \code{\link[nloptr]{nloptr}}.}

\item{start.b.fraction}{The estamated fraction of the mutations in
\code{spectra} due to the background signature.}
}
\description{
We index mutation channels (e.g. \code{ACA > AAA, ACC > AAC, ...}) by \eqn{j}, \eqn{j \in 1...96}.

We index input mutational spectra by \eqn{i}. 

Let
 
\eqn{g = g_1, g_2,\ldots , g_{96}}, with \eqn{\Sigma_j g_j = 1}
 be the previously determined, input background signature profile,

\eqn{s^i, i \in 1, 2,\ldots} be the input spectra,
from exposed samples, usually only 2 or 3,

\eqn{b^i, i \in 1, 2,\ldots} be the (to-be-estimated)
numbers of mutations due to the background signature in each
\eqn{s^i}, and

\eqn{t = t_1, t_2,\ldots , t_{96}}, with \eqn{\Sigma_j t_j = 1}
be the (to-be-estimated) target signature due to an exposure.

We want to maximize \eqn{\Pi^iP(s^i|b^i,t)P(b^i)} over 
\eqn{\vec{b} = b^1, b^2,\dots} and \eqn{t}.

\eqn{P(b^i)} is estimated from the distribution of previously observed
numbers of mutations in untreated samples.

\eqn{P(s^i|b^i,t)} is estimated as follows:

Let \eqn{|s^i|} denote
the total number of mutations in spectrum \eqn{s^i}, i.e.
\eqn{|s^i| = \Sigma_j s^i_j}, \eqn{j \in 1...96}.

The expected number of mutations in each
mutation category, \eqn{j}, is estimated as

\eqn{e^i_j = g_jb_j^i + t_j(|s^i| - b^i)}.

Then \eqn{P(s^i|e^i)} is estimated as \eqn{\Pi_jP(s^i_j|e^i_j)} 

given a negative binomial distribution
centered on each \eqn{e^i_j}, all distributions
with the same dispersion parameter
estimated as described
below (need to describe / review).
}
\details{
See \code{\link{ObjFn1}}.
}
