% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetDistanceInfo}
\alias{GetDistanceInfo}
\title{Retrieve distance information from the output generated by running
\code{MAPAssignActivity1} or \code{SparseAssignActivity1} on multiple samples}
\usage{
GetDistanceInfo(
  list.of.MAP.out,
  sparse.assign = FALSE,
  use.forward.search = FALSE
)
}
\description{
Retrieve distance information from the output generated by running
\code{MAPAssignActivity1} or \code{SparseAssignActivity1} on multiple samples
}
\keyword{internal}
