% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{GetExposureInfo}
\alias{GetExposureInfo}
\title{Retrieve exposure information from the output generated by \code{MAPAssignActivity}}
\usage{
GetExposureInfo(list.of.MAP.out)
}
\description{
Retrieve exposure information from the output generated by \code{MAPAssignActivity}
}
\keyword{internal}
