% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HepG2.background.spectra}
\alias{HepG2.background.spectra}
\title{Test background spectra}
\format{XXXX}
\usage{
HepG2.background.spectra
}
\description{
Test background spectra
}
\examples{
# XXXX
}
\keyword{datasets}
