% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.sigs}
\alias{kucab.sigs}
\title{Kucab et al., 2019 signatures as a standard \code{ICAMS} counts.signature catalog.}
\format{An object of class \code{SBS96Catalog} (inherits from \code{matrix}) with 96 rows and 54 columns.}
\source{
https://github.com/xqzou/Cell_MutagenSig
}
\usage{
kucab.sigs
}
\description{
Kucab et al., 2019 signatures as a standard \code{ICAMS} counts.signature catalog.
}
\keyword{datasets}
