% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kucab.sub.catalog}
\alias{kucab.sub.catalog}
\title{Kucab spectra -- 2 mutation label columns and 324 data columns.}
\format{An object of class \code{data.frame} with 96 rows and 326 columns.}
\source{
https://github.com/xqzou/Cell_MutagenSig
}
\usage{
kucab.sub.catalog
}
\description{
The original format used in Xueqing ZOU's code.
}
\keyword{datasets}
