% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubtractBackground.R
\name{LLHOfSignatureGivenSpectrum}
\alias{LLHOfSignatureGivenSpectrum}
\title{The log likelihood of a \strong{single} observed spectrum given the signature.}
\usage{
LLHOfSignatureGivenSpectrum(spectrum, signature, nbinom.size)
}
\arguments{
\item{spectrum}{A single background spectrum.}

\item{signature}{A signature as a numeric vector.}

\item{nbinom.size}{The \code{size} argument for
\code{\link[stats]{NegBinomial}}.}
}
\description{
Assumes that all mutations \code{spectrum} were generated by \code{signature}.
}
