% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparePackageData.R
\name{MakeCisplatinCatalogs}
\alias{MakeCisplatinCatalogs}
\title{Make spectrum catalog from VCFs from cisplatin exposed HepG2}
\usage{
MakeCisplatinCatalogs()
}
\value{
The catalog
}
\description{
Make spectrum catalog from VCFs from cisplatin exposed HepG2
}
\keyword{internal}
