% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeSynBackground}
\alias{MakeSynBackground}
\title{Generate \code{num.samples} synthetic background spectra.}
\usage{
MakeSynBackground(background.info, num.samples)
}
\arguments{
\item{background.info}{See for example \code{HepG2.background.info}.}

\item{num.samples}{Number of synthetic spectra to generate.}
}
\description{
Generate \code{num.samples} synthetic background spectra.
}
