% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeSynBackground}
\alias{MakeSynBackground}
\title{Generate \code{num.samples} synthetic background spectra.}
\usage{
MakeSynBackground(background.info, num.samples,
  total.count.nbinom.size = NULL, sig.nbinom.size = NULL)
}
\arguments{
\item{background.info}{See for example \code{HepG2.background.info}.}

\item{num.samples}{Number of synthetic spectra to generate.}

\item{total.count.nbinom.size}{If not \code{NULL} then
override the \code{count.nbinom.size}
parameter in \code{background.info}.}

\item{sig.nbinom.size}{If note \code{NULL} the override
the \code{sig.nbinom.size} parameter in in \code{background.info}.}
}
\description{
Generate \code{num.samples} synthetic background spectra.
}
