% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynBGSignatureTests.R
\name{MakeTestSpectrum}
\alias{MakeTestSpectrum}
\title{Make a single test spectrum.}
\usage{
MakeTestSpectrum(
  replicate,
  bg.sig.info,
  bg.contribution,
  target.sig.name = NULL,
  ref.sig = NULL
)
}
\arguments{
\item{replicate}{Arbitrary index for this test}

\item{bg.sig.info}{See, for example, \code{HepG2.background.info}.}

\item{bg.contribution}{Expected proportion of mutations due to the background.}

\item{target.sig.name}{Name of a standard SigProfiler PCAWG7 signature.}

\item{ref.sig}{Target signature as a signature, rather than string}
}
\description{
Make a single test spectrum.
}
