% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{MAPAssignActivity}
\alias{MAPAssignActivity}
\title{Find Maximum A Posteriori (MAP) assignment of signature exposures that
explain multiple spectra}
\usage{
MAPAssignActivity(
  spectra,
  sigs,
  sigs.presence.prop,
  output.dir,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  num.parallel.samples = 5,
  mc.cores.per.sample = min(20, 2^max.level),
  max.subsets = 1000,
  max.presence.proportion = 0.99,
  progress.monitor = NULL,
  seed = NULL
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}.}

\item{output.dir}{Directory path to save the output file.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with as opposed to
without a set of signatures
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{num.parallel.samples}{The (maximum) number of samples to run in
parallel. On Microsoft Windows machines it is silently changed to 1. Each
sample in turn can require multiple cores, as governed by
\code{mc.cores.per.sample}.}

\item{mc.cores.per.sample}{The maximum number of cores to use for each
sample. On Microsoft Windows machines it is silently changed to 1.}

\item{max.subsets}{The maximum number of subsets that can be
tested for removal from the set of signatures.}

\item{max.presence.proportion}{The maximum value of the proportion
of tumors that must have a given signature.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}
}
\value{
A list of lists containing output for each sample in \code{spectra}. 
Each sublist has the following elements \describe{

\item{MAP}{A 2-column \code{tibble} with the attributions with the highest MAP found.
   Column 1 contains signature ids; column 2 contains the associated counts. }

\item{MAP.row}{A 1-row \code{tibble} with various information on the selected exposure.}

\item{best.sparse}{A 2-column \code{tibble} with the most-sparse attributions with
     the highest MAP, in the same format as element \code{MAP}.}

\item{best.sparse.row}{{A 1-row \code{tibble} with various information on the
   most-sparse exposure with the best MAP.}}

\item{all.tested}{A \code{tibble} of all the search results.}

\item{messages}{Possibly empty character vector with messages.}

\item{success}{\code{TRUE} is search was successful, \code{FALSE} otherwise.}

\item{time.for.MAP.assign}{Value from \code{system.time} for
 \code{\link{MAPAssignActivityInternal}}}.

\item{MAP.recon}{Reconstruction based on \code{MAP}}.

\item{sparse.MAP.recon}{Reconstruction based on \code{best.sparse}}.

\item{MAP.distances}{Various distances and similarities
between \code{spect} and \code{MAP.recon}}.

\item{sparse.MAP.distances}{Various distances and similarities
between \code{spect} and \code{sparse.MAP.recon}}.

}

These elements will be \code{NULL} if \code{max.subsets} is exceeded.
}
\description{
Find Maximum A Posteriori (MAP) assignment of signature exposures that
explain multiple spectra
}
