% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{MAPAssignActivity}
\alias{MAPAssignActivity}
\title{Find Maximum A Posteriori (MAP) assignment of signature exposures that
explain multiple spectra}
\usage{
MAPAssignActivity(
  spectra,
  sigs,
  sigs.presence.prop,
  output.dir,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  num.parallel.samples = 5,
  mc.cores.per.sample = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{output.dir}{Directory path to save the output file.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with as opposed to
without a set of signatures
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{num.parallel.samples}{The (maximum) number of samples to run in
parallel. On Microsoft Windows machines it is silently changed to 1. Each
sample in turn can require multiple cores, as governed by
\code{mc.cores.per.sample}.}

\item{mc.cores.per.sample}{The maximum number of cores to use for each
sample. On Microsoft Windows machines it is silently changed to 1.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: Proposed signature assignment for \code{spectra}
with the highest MAP found.
\item \code{proposed.reconstruction}: Proposed reconstruction of \code{spectra} based on \code{MAP}.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spectra} and \code{proposed.reconstruction}.
\item \code{error.messages}: Only appearing if there are errors running
\code{MAPAssignActivity}.
\item \code{results.details}: Detailed results for each sample in \code{spectra}.
}

These elements will be \code{NULL} if the algorithm could not find the
optimal reconstruction or there are errors coming out.
}
\description{
Find Maximum A Posteriori (MAP) assignment of signature exposures that
explain multiple spectra
}
\examples{
\dontrun{
# This is a long running example unless multiple CPU cores are available
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
sigs <- PCAWG7::signature$genome$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
MAP.out <- MAPAssignActivity(spectra = spectra,
                             sigs = sigs,
                             sigs.presence.prop = sigs.prop,
                             output.dir = file.path(tempdir(), "Lung-AdenoCA"),
                             max.level = length(sigs.prop) - 1,
                             p.thresh = 0.05 / ncol(spectra),
                             num.parallel.samples = 2,
                             mc.cores.per.sample = 10)
}
}
