% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{MAPAssignActivity1}
\alias{MAPAssignActivity1}
\title{Find a Maximum A Posteriori assignment of signature exposures for one spectrum.}
\usage{
MAPAssignActivity1(
  spect,
  sigs,
  sigs.presence.prop,
  max.level = 5,
  p.thresh = 0.05,
  eval_f,
  m.opts,
  max.mc.cores = min(20, 2^max.level),
  max.subsets = 1000,
  max.presence.proportion = 0.99
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being the same as \code{colnames(sigs)}.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with than without a set of signatures
is > than \code{p.thresh}, then we can use exposures without this set.}

\item{eval_f}{See \code{\link[nloptr]{nloptr}}.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{max.subsets}{The maxium number of subsets that can be
tested for removal from the set of signatures.}

\item{max.presence.proportion}{The maxium value of the proportion
of tumors that must have a given signature.}
}
\value{
A list with the elements \describe{

\item{MAP}{A 1-column tibble with the attributions with the highest MAP found.}

\item{most.sparse}{A 1-column tibble with the most-sparse attributions with the highest MAP.}

\item{all.tested}{A tibble of all the search results.}

\item{messages}{Possibly empty character vector with messages.}

\item{success}{\code{TRUE} is search was successful, \code{FALSE} otherwise.}

}

These elements will be \code{NULL} if \code{max.subsets} is exceeded.
}
\description{
Find a Maximum A Posteriori assignment of signature exposures for one spectrum.
}
\keyword{internal}
