% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{MAPAssignActivity1}
\alias{MAPAssignActivity1}
\title{Find a Maximum A Posteriori (MAP) assignment of signature exposures that explain one spectrum.}
\usage{
MAPAssignActivity1(
  spect,
  sigs,
  sigs.presence.prop,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  max.mc.cores = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL,
  max.subsets = 1000,
  use.sparse.assign = FALSE,
  drop.low.mut.samples = TRUE
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with a
set of signatures (as opposed to
without that set of signatures)
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}

\item{max.subsets}{This argument provides a way to 
heuristically limit the
amount of time spent by this function. Larger values of this
argument will tend to allow longer running times.
The algorithm
successively tries to remove all subsets of 1 signature, 2
signatures, 3 signatures, etc., down to \code{max.level}.
(Not every subset is tested at each level; if a subset was
already found to be necessary the algorithm does not test
supersets of that subset.) If at any level the algorithm
needs to test more than \code{max.subsets} this function will
not proceed.}

\item{use.sparse.assign}{Whether to use sparse assignment. If \code{TRUE},
arguments designed for Maximum A Posteriori assignment such as
\code{sigs.presence.prop} will be ignored.}

\item{drop.low.mut.samples}{Whether to exclude low mutation samples from
the analysis. If \code{TRUE(default)}, samples with SBS total mutations less
than 100, DBS or ID total mutations less than 25 will be dropped.}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: Proposed signature assignment for \code{spect}
with the highest MAP found. If \code{use.sparse.assign = TRUE}, this will
be the most sparse set of signatures that can plausibly explain \code{spect}.
\item \code{proposed.reconstruction} :Reconstruction based on \code{MAP}.
If \code{use.sparse.assign = TRUE}, this will be the reconstruction based on
sparse assignment.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spect} and \code{proposed.reconstruction}.
\item \code{all.tested}: A \code{tibble} of all the search results.
\item \code{alt.solutions}: A \code{tibble} showing all the alternative solutions
that are statistically as good as the \code{proposed.assignment} that can
plausibly reconstruct \code{spect}.
\item \code{time.for.MAP.assign}: Value from \code{system.time} for running
\code{MAPAssignActivity1}.
\item \code{error.messages}: Only present if there were errors running
\code{MAPAssignActivity1}.
}

The elements \code{proposed.assignment}, \code{proposed.reconstruction},
\code{reconstruction.distances}, \code{all.tested},
\code{time.for.MAP.assign} will be \code{NULL} if the algorithm could not
find the optimal reconstruction or there are errors coming out.
}
\description{
This function also can do sparse assignment by specifying \code{use.sparse.assign = TRUE}.
}
\examples{
\dontrun{
# This is a long running example unless parallel computing is supported on your machine
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spect <- PCAWG7::spectra$PCAWG$SBS96[, indices[1], drop = FALSE]
sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96", 
                                 cancer.type = "Lung-AdenoCA")
MAP.out <- MAPAssignActivity1(spect = spect, 
                              sigs = sigs, 
                              sigs.presence.prop = sigs.prop, 
                              max.level = length(sigs.prop) - 1)
}                               
}
\keyword{internal}
