% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{MAPAssignActivity1}
\alias{MAPAssignActivity1}
\title{Find a Maximum A Posteriori (MAP) assignment of signature exposures that explain one spectrum.}
\usage{
MAPAssignActivity1(
  spect,
  sigs,
  sigs.presence.prop,
  max.level = 5,
  p.thresh = 0.05,
  eval_f,
  eval_g_ineq = NULL,
  m.opts,
  max.mc.cores = min(20, 2^max.level),
  max.subsets = 1000,
  max.presence.proportion = 0.99,
  progress.monitor = NULL
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with than without a set of signatures
is > than \code{p.thresh}, then we can use exposures without this set.}

\item{eval_f}{See \code{\link[nloptr]{nloptr}}.}

\item{eval_g_ineq}{See \code{\link[nloptr]{nloptr}}.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{max.subsets}{The maximum number of subsets that can be
tested for removal from the set of signatures.}

\item{max.presence.proportion}{The maximum value of the proportion
of tumors that must have a given signature.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}
}
\value{
A list with the elements \describe{

\item{MAP}{A 2-column \code{tibble} with the attributions with the highest MAP found.
   Column 1 contains signature ids; column 2 contains the associated counts. }

\item{MAP.row}{A 1-row \code{tibble} with various information on the selected exposure.}

\item{best.sparse}{A 2-column \code{tibble} with the most-sparse attributions with
     the highest MAP, in the same format as element \code{MAP}.}

\item{best.sparse.row}{{A 1-row \code{tibble} with various information on the
   most-sparse exposure with the best MAP.}}

\item{all.tested}{A \code{tibble} of all the search results.}

\item{messages}{Possibly empty character vector with messages.}

\item{success}{\code{TRUE} is search was successful, \code{FALSE} otherwise.}

\item{time.for.MAP.assign}{Value from \code{system.time} for
 \code{\link{MAPAssignActivityInternal}}}.

\item{MAP.recon}{Reconstruction based on \code{MAP}}.

\item{sparse.MAP.recon}{Reconstruction based on \code{best.sparse}}.

\item{MAP.distances}{Various distances and similarities
between \code{spect} and \code{MAP.recon}}.

\item{sparse.MAP.distances}{Various distances and similarities
between \code{spect} and \code{sparse.MAP.recon}}.

}

These elements will be \code{NULL} if \code{max.subsets} is exceeded.
}
\description{
Find a Maximum A Posteriori (MAP) assignment of signature exposures that explain one spectrum.
}
