% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{MAPAssignActivity1}
\alias{MAPAssignActivity1}
\title{Component of \code{\link{SparseAssignActivity}} for one spectrum.}
\usage{
MAPAssignActivity1(
  spect,
  sigs,
  sigs.presence.prop,
  max.level = 5,
  p.thresh = 0.05,
  eval_f,
  m.opts,
  max.mc.cores = min(20, 2^max.level),
  max.subsets = 1000
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being the same as \code{colnames(sigs)}.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{The p value threshold for deciding if a set of signatures is necessary.}

\item{eval_f}{See \code{\link[nloptr]{nloptr}}.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.)}
}
\description{
Component of \code{\link{SparseAssignActivity}} for one spectrum.
}
