% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1.R
\name{MAPAssignActivity1}
\alias{MAPAssignActivity1}
\title{Find a Maximum A Posteriori (MAP) assignment of signature exposures that explain one spectrum.}
\usage{
MAPAssignActivity1(
  spect,
  sigs,
  sigs.presence.prop,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  max.mc.cores = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{sigs.presence.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being a subset of \code{colnames(sigs)}. See \code{\link{ExposureProportions}}
for more details.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with as opposed to
without a set of signatures
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: Proposed signature assignment for \code{spect}
with the highest MAP found.
\item \code{proposed.reconstruction} :Reconstruction based on \code{MAP}.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spect} and \code{proposed.reconstruction}.
\item \code{all.tested}: A \code{tibble} of all the search results.
\item \code{time.for.MAP.assign}: Value from \code{system.time} for running
\code{MAPAssignActivity1}.
\item \code{error.messages}: Only appearing if there are errors running
\code{MAPAssignActivity1}.
}

These elements will be \code{NULL} if the algorithm could not find the
optimal reconstruction or there are errors coming out.
}
\description{
Find a Maximum A Posteriori (MAP) assignment of signature exposures that explain one spectrum.
}
\examples{
\dontrun{
# This is a long running example unless multiple CPU cores are available
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spect <- PCAWG7::spectra$PCAWG$SBS96[, indices[1], drop = FALSE]
sigs <- PCAWG7::signature$genome$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96", 
                                 cancer.type = "Lung-AdenoCA")
MAP.out <- MAPAssignActivity1(spect = spect, 
                              sigs = sigs, 
                              sigs.presence.prop = sigs.prop, 
                              max.level = length(sigs.prop) - 1)
}                               
}
