% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nloptr1Tumor.R
\name{Nloptr1Tumor}
\alias{Nloptr1Tumor}
\title{Optimize assignment of a fixed set of signature activities for a \strong{single} tumor.}
\usage{
Nloptr1Tumor(spectrum, sigs, m.opts = NULL, ...)
}
\arguments{
\item{spectrum}{The spectrum to be reconstructed.}

\item{sigs}{The available signatures.}

\item{m.opts}{Options that govern the numerical optimization.
For documentation see \code{\link{DefaultManyOpts}}.}

\item{...}{Additional arguments for \code{eval_f}.}
}
\description{
Optimize assignment of a fixed set of signature activities for a \strong{single} tumor.
}
\keyword{internal}
