% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{obj.fun.nbinom.maxlh}
\alias{obj.fun.nbinom.maxlh}
\title{Old objective function for SparseAssignActivity}
\usage{
obj.fun.nbinom.maxlh(exp, spectrum, sigs, nbinom.size)
}
\description{
Old objective function for SparseAssignActivity
}
\keyword{internal}
