% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubtractBackground.R
\name{ObjFn1}
\alias{ObjFn1}
\title{Objective function for \code{\link{FindSignatureMinusBackground}}.}
\usage{
ObjFn1(est.target.sig.and.b, obs.spectra, bg.sig.info)
}
\arguments{
\item{est.target.sig.and.b}{A numeric vector of which elements
1:96 are the signature (as a vector) and the remaining elements
are the coefficients for each input spectrum in \code{obs.spectra}.}

\item{obs.spectra}{The observed spectra, which may be the
sum of the activities of the background signature and the signature
of an experimental treatment.}

\item{bg.sig.info}{Information on the background signature. See for example
\code{\link{HepG2.background.info}}.

We the caller will seek to minimize the value of this function.}
}
\description{
Objective function for \code{\link{FindSignatureMinusBackground}}.
}
