% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{ObjFnBinomMaxLH2}
\alias{ObjFnBinomMaxLH2}
\title{The negative binomial maximum likelihood objective function.}
\usage{
ObjFnBinomMaxLH2(
  exp,
  spectrum,
  sigs,
  nbinom.size,
  no.round.ok = FALSE,
  show.warning = FALSE
)
}
\arguments{
\item{exp}{The matrix of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}

\item{nbinom.size}{The dispersion parameter for the negative
binomial distribution; smaller is more dispersed.
See \code{\link[stats]{NegBinomial}}.}

\item{no.round.ok}{If \code{TRUE}, allow use of unrounded
reconstruction if some mutation types would have 0
counts in the reconstructed spectrum.}

\item{show.warning}{If \code{TRUE} print warning if unrounded
reconstructions were used.}
}
\description{
For use by \code{\link[nloptr]{nloptr}}
}
\keyword{internal}
