% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjFnDMMaxLH.R
\name{ObjFnDMMaxLH}
\alias{ObjFnDMMaxLH}
\title{The Dirichlet-multinomial likelihood objective function}
\usage{
ObjFnDMMaxLH(exp, spectrum, sigs, num.replicates = 1000)
}
\arguments{
\item{exp}{The matrix of exposures ("activities").}

\item{spectrum}{The spectrum to assess.}

\item{sigs}{The matrix of signatures.}

\item{num.replicates}{Number of bootstrap replicates for reconstructed spectrum.}
}
\value{
-1 * log(likelihood(spectrum | reconstruction))

\code{\link[nloptr]{nloptr}} minimizes the objective function, so the
lower the objective function, the better.
}
\description{
Can be used as the
objective function for \code{\link{MAPAssignActivity}},
\code{\link{MAPAssignActivity1}},
\code{\link{SparseAssignActivity}},
\code{\link{SparseAssignActivity1}},
\code{\link{SignaturePresenceTest}}
and \code{\link{SignaturePresenceTest1}}.
(Internally used by by \code{\link[nloptr]{nloptr}}.)
}
\keyword{internal}
