% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{one.lh.and.exp}
\alias{one.lh.and.exp}
\title{Returns a list with elements:}
\usage{
one.lh.and.exp(spect, sigs, m.opts, eval_f)
}
\description{
\code{loglh}    - the log likelihood of the best solution (set of exposures) found
\code{exposure}  - the vector of exposures that generate \code{loglh}, in this case
this is the number of mutations ascribed to each signature.
}
\keyword{internal}
