% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMAPAssignTest.R
\name{OneMAPAssignTest}
\alias{OneMAPAssignTest}
\title{Run one test of MAPAssignActivity1}
\usage{
OneMAPAssignTest(
  spect,
  reference.exp,
  cancer.type,
  mutation.type,
  max.subsets = 1000,
  max.level = 5,
  max.mc.cores = 100
)
}
\description{
Run one test of MAPAssignActivity1
}
