% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{OptimizeExposure}
\alias{OptimizeExposure}
\title{Optimize the reconstruction of a spectrum from a set of signatures.}
\usage{
OptimizeExposure(spect, sigs, m.opts, eval_f, ...)
}
\arguments{
\item{spect}{The spectrum to be reconsructed.}

\item{sigs}{The available signatures.}

\item{m.opts}{Options that govern the numerical optimizaiton. 
For documentation see \code{\link{DefaultManyOpts}}.}

\item{eval_f}{The objective function for
\code{\link[nloptr]{nloptr}}. We have only tested
\code{\link{ObjFnBinomMaxLHNoRoundOK}} and
\code{\link{ObjFnBinomMaxLHMustRound}}.}

\item{...}{Additional arguments for \code{eval_f}.

Returns a list with elements：

\code{loglh}    - the log likelihood of the best solution (set of exposures) found
\code{exposure}  - the vector of exposures that generate \code{loglh}, in this case
this is the number of mutations ascribed to each signature.}
}
\description{
Optimize the reconstruction of a spectrum from a set of signatures.
}
\keyword{export}
