% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{OptimizeExposure}
\alias{OptimizeExposure}
\title{Optimize the reconstruction of a spectrum from a set of signatures.}
\usage{
OptimizeExposure(spectrum, sigs, m.opts, eval_f, ...)
}
\arguments{
\item{spectrum}{The spectrum to be reconsructed.}

\item{sigs}{The available signatures.}

\item{m.opts}{Options that govern the numerical optimizaiton. 
For documentation see \code{\link{DefaultManyOpts}}.}

\item{eval_f}{The objective function for
\code{\link[nloptr]{nloptr}}. We have only tested
\code{\link{ObjFnBinomMaxLHNoRoundOK}} and
\code{\link{ObjFnBinomMaxLHMustRound}}.}

\item{...}{Additional arguments for \code{eval_f}.

Returns a list with elements \describe{
\item{\code{loglh}}{The log likelihood of the best solution (set of exposures) found.
      For a more general objective function this might be \code{NULL}.}
\item{\code{exposure}}{The vector of exposures that generate \code{loglh}, i.e. 
   the number of mutations ascribed to each signature.}
\item{\code{obj.fn.value}}{The objective function value associated with \code{exposure}.}
\item{\code{everything.else}}{Everything returned by the function \code{\link{Nloptr1Tumor}.} }
}}
}
\description{
Optimize the reconstruction of a spectrum from a set of signatures.
}
