% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeExposureQP.R
\name{OptimizeExposureQPBootstrap}
\alias{OptimizeExposureQPBootstrap}
\title{Bootstrap \code{\link{OptimizeExposureQP}} and filter exposures by confidence intervals}
\usage{
OptimizeExposureQPBootstrap(
  spectrum,
  signatures,
  num.replicates = 10000,
  conf.int = 0.95,
  mc.cores = 10
)
}
\arguments{
\item{spectrum}{Mutational signature spectrum
as a numeric vector or single column data frame or matrix.}

\item{signatures}{Matrix or data frame of signatures from which
reconstruct \code{spectrum}. Rows are mutation types and
columns are signatures. Should have column names for
interpretable results. Cannot be a vector because
the column names are needed.}

\item{num.replicates}{Number of bootstrap replicates.}

\item{mc.cores}{The maximum number of cores to use.
On MS Windows machines it defaults to 1.}

\item{conf.interval}{Discard signatures with \code{conf.int} that overlaps 0.}
}
\description{
Bootstrap \code{\link{OptimizeExposureQP}} and filter exposures by confidence intervals
}
