% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMAPAssignTest.R
\name{PCAWGMAPTest}
\alias{PCAWGMAPTest}
\title{Run \code{\link{MAPAssignActivity1}} on one sample from the PCAWG platinum data set.}
\usage{
PCAWGMAPTest(
  cancer.type,
  sample.index,
  mutation.type,
  max.level = 5,
  max.mc.cores,
  out.dir = NULL,
  p.thresh = 0.01,
  m.opts = DefaultManyOpts(),
  max.presence.proportion = 0.99
)
}
\arguments{
\item{cancer.type}{A cancer type from the PCAWG exposures matrix.}

\item{sample.index}{The index of the sample within the exposures matrix.}

\item{mutation.type}{One of "SBS96", "SBS192", "ID", "DBS78"}

\item{max.level}{The maximum number of signatures to try removing.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{out.dir}{If non-NULL creat this directory if necessary and put
results there.}

\item{p.thresh}{If
the p value for a better reconstruction with than without a set of signatures
is > than \code{p.thresh}, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.presence.proportion}{xxxxx}
}
\description{
Run \code{\link{MAPAssignActivity1}} on one sample from the PCAWG platinum data set.
}
