% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity1Test.R, R/PCAWGMAPTest.R
\name{PCAWGMAPTest}
\alias{PCAWGMAPTest}
\title{Run \code{\link{MAPAssignActivity1}} on one sample from the PCAWG platinum data set.}
\usage{
PCAWGMAPTest(
  cancer.type,
  sample.index,
  mutation.type,
  max.level = 5,
  max.mc.cores,
  out.dir = NULL,
  p.thresh = 0.01,
  m.opts = DefaultManyOpts(),
  max.presence.proportion = 0.99,
  sigs.prop = NULL
)

PCAWGMAPTest(
  cancer.type,
  sample.index,
  mutation.type,
  max.level = 5,
  max.mc.cores,
  out.dir = NULL,
  p.thresh = 0.01,
  m.opts = DefaultManyOpts(),
  max.presence.proportion = 0.99,
  sigs.prop = NULL
)
}
\arguments{
\item{cancer.type}{A cancer type from the PCAWG exposures matrix.}

\item{sample.index}{The index of the sample within the exposures matrix.}

\item{mutation.type}{One of "SBS96", "SBS192", "ID", "DBS78"}

\item{max.level}{The maximum number of signatures to try removing.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.}

\item{out.dir}{If non-NULL create this directory if necessary and put
results there.}

\item{p.thresh}{If
the p value for a better reconstruction with than without a set of signatures
is > than \code{p.thresh}, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.presence.proportion}{The maximum value of the proportion
of tumors that must have a given signature. Used so that it is
possible to exclude a signature from a spectrum, e.g.
perhaps all examples of tumor types have SBS5, but we want
to allow a small chance that SBS5 is not present.}

\item{sigs.prop}{The proportions of samples that contain each
signature. A numerical vector (values between 0 and 1), with names
being signature identifiers. Can be the
return value from \code{\link{ExposureProportions}}.}
}
\value{
See \code{\link{OneMAPAssignTest}}.

A list with two elements, each the result for one call to \code{\link{OneMAPAssignTest}}.
}
\description{
Run \code{\link{MAPAssignActivity1}} on one sample from the PCAWG platinum data set.

Run \code{\link{MAPAssignActivity1}} on one sample from the PCAWG platinum data set with artifact signatures removed.
}
\keyword{internal}
