% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMAPAssignTest.R
\name{PCAWGMAPTest}
\alias{PCAWGMAPTest}
\title{Run \code{\link{MAPAssignSignature1}} on one sample from the PCAWG platinum data set.}
\usage{
PCAWGMAPTest(
  cancer.type,
  sample.index,
  mutation.type,
  max.level = 5,
  max.mc.cores,
  out.dir = NULL,
  p.thresh = 0.01
)
}
\arguments{
\item{cancer.type}{A cancer type from the exposures matrix.}

\item{sample.index}{The index of the sample within the exposures matrix.}

\item{mutation.type}{One of "SBS96", "SBS192", "ID", "DBS78"}
}
\description{
Run \code{\link{MAPAssignSignature1}} on one sample from the PCAWG platinum data set.
}
