% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPAssignActivity.R
\name{PlotListOfCatalogsToPdf}
\alias{PlotListOfCatalogsToPdf}
\title{Plot List of catalogs to Pdf}
\usage{
PlotListOfCatalogsToPdf(
  list.of.catalogs,
  file,
  plot.SBS12 = FALSE,
  cex = 0.8,
  grid = TRUE,
  upper = TRUE,
  xlabels = TRUE,
  ylim = NULL
)
}
\arguments{
\item{list.of.catalogs}{List of catalogs in \code{\link{ICAMS}} format.}

\item{file}{The name of the PDF file to be produced.}

\item{plot.SBS12}{Only meaningful for class \code{SBS192Catalog}; if \code{TRUE},
generate an abbreviated plot of only SBS without context, i.e. 
C>A, C>G, C>T, T>A, T>C, T>G each on transcribed and untranscribed strands,
rather than SBS in trinucleotide context, e.g.
ACA > AAA, ACA > AGA, ..., TCT > TAT, ... There are 12 bars in the graph.}

\item{cex}{Has the usual meaning. Taken from \code{par("cex")} by default.
Only implemented for SBS96Catalog, SBS192Catalog and DBS144Catalog.}

\item{grid}{A logical value indicating whether to draw grid lines. Only
implemented for SBS96Catalog, DBS78Catalog, IndelCatalog, ID166Catalog.}

\item{upper}{A logical value indicating whether to draw horizontal lines and
the names of major mutation class on top of graph. Only implemented for
SBS96Catalog, DBS78Catalog, IndelCatalog, ID166Catalog.}

\item{xlabels}{A logical value indicating whether to draw x axis labels. Only
implemented for SBS96Catalog, DBS78Catalog, IndelCatalog, ID166Catalog. 
If \code{FALSE} then plot x axis tick marks for SBS96Catalog; set
\code{par(tck = 0)} to suppress.}

\item{ylim}{Has the usual meaning. Only implemented for SBS96Catalog,
IndelCatalog, ID166Catalog.}
}
\description{
Plot List of catalogs to Pdf
}
\keyword{internal}
