% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAssignActivity.R
\name{PresenceAssignActivity}
\alias{PresenceAssignActivity}
\title{Find minimal set of signatures that can explain multiple spectra by first using
signature presence test}
\usage{
PresenceAssignActivity(
  spectra,
  sigs,
  output.dir,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = DefaultManyOpts(),
  num.parallel.samples = 5,
  mc.cores.per.sample = min(20, 2^max.level),
  progress.monitor = NULL,
  seed = NULL,
  drop.low.mut.samples = TRUE,
  use.forward.search = FALSE,
  sig.pres.test.nbinom.size = NULL,
  sig.pres.test.p.thresh = 0.05,
  save.files = TRUE
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{output.dir}{Directory path to save the output file.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{If
the p value for a better reconstruction with a
set of signatures (as opposed to
without that set of signatures)
is > than this argument, then we can use exposures without this set.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{num.parallel.samples}{The (maximum) number of samples to run in
parallel. On Microsoft Windows machines it is silently changed to 1. Each
sample in turn can require multiple cores, as governed by
\code{mc.cores.per.sample}.}

\item{mc.cores.per.sample}{The maximum number of cores to use for each
sample. On Microsoft Windows machines it is silently changed to 1.}

\item{progress.monitor}{Function called at the start of each
new level (number of signatures to try excluding). Must
take named arguments \code{value} and \code{detail}, and
no others. Designed for a \code{\link[ipc]{AsyncProgress}}
progress bar function.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}

\item{drop.low.mut.samples}{Whether to exclude low mutation samples from
the analysis. If \code{TRUE}, samples with SBS total mutations less
than 100, DBS or ID total mutations less than 25 will be dropped.}

\item{use.forward.search}{Whether to use forward search to find the minimal
number of signatures to optimally reconstruct the spectrum.}

\item{sig.pres.test.nbinom.size}{The dispersion parameter for the negative
binomial distribution used when conducting signature presence test first to
filter out those signatures that are not needed in the reconstruction of
the spectrum; smaller is more dispersed. See
\code{\link[stats]{NegBinomial}}. If \code{NULL}, then use multinomial
likelihood when conducting signature presence test.}

\item{sig.pres.test.p.thresh}{If the p-value of signature presence test for
one signature is >= \code{sig.pres.test.p.thresh}, then this signature will
not be used for assignment later.}

\item{save.files}{If \code{TRUE} save several files for each input sample
in a directory named after the sample.}
}
\value{
A list with the elements:
\itemize{
\item \code{proposed.assignment}: The proposed set of signatures that can
plausibly explain \code{spectra}.
\item \code{proposed.reconstruction}: The reconstruction based on
\code{proposed.assignment}.
\item \code{reconstruction.distances}: Various distances and similarities
between \code{spectra} and \code{proposed.reconstruction}.
\item \code{all.tested}: All tested possible ways to reconstruct each
sample in \code{spectra}.
\item \code{alt.solutions}: A \code{tibble} showing all the alternative solutions
that are statistically as good as the \code{proposed.assignment} that can
plausibly reconstruct \code{spectra}.
\item \code{time.for.assignment}: Value from \code{system.time} for running
\code{PresenceAssignActivity} for each sample in \code{spectra}.
\item \code{error.messages}: Only appearing if there are errors running
\code{PresenceAssignActivity}.
}

The elements \code{proposed.assignment}, \code{proposed.reconstruction},
\code{reconstruction.distances}, \code{all.tested},
\code{time.for.assignment} will be \code{NULL} if the algorithm could not
find the optimal reconstruction or there are errors coming out for
\strong{all} samples.
}
\description{
Find minimal set of signatures that can explain multiple spectra by first using
signature presence test
}
\examples{
\dontrun{
# This is a long running example unless parallel computing is supported on your machine
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
SBS96.sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
sigs.to.use <- SBS96.sigs[, names(sigs.prop), drop = FALSE]
retval <- PresenceAssignActivity(spectra = spectra,
                                 sigs = sigs.to.use,
                                 output.dir = file.path(tempdir(), "Lung-AdenoCA"),
                                 max.level = ncol(sigs.to.use) - 1,
                                 p.thresh = 0.05 / ncol(sigs.to.use),
                                 num.parallel.samples = 2,
                                 mc.cores.per.sample = 30,
                                 seed = 2561)
}
}
