% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ReconstructSpectrum}
\alias{ReconstructSpectrum}
\title{Given signatures (sigs) and exposures (exp), return a spectrum or spectra}
\usage{
ReconstructSpectrum(sigs, exp, use.sig.names = FALSE)
}
\arguments{
\item{sigs}{Signature as a matrix or data frame, with each
row one mutation type (e.g. CCT > CAT or CC > TT) and
each column a signature.}

\item{exp}{The exposures for one or more samples as a matrix or data.frame,
with each row a signature and each column a sample.}

\item{use.sig.names}{If \code{TRUE} check that \code{rownames(exp)} is
a subset of \code{colnames(sigs)}, and use only the columns in \code{sigs}
that are present in \code{exp}.}
}
\value{
The matrix product \code{sigs \%*\% exp} after some error checking.
}
\description{
Given signatures (sigs) and exposures (exp), return a spectrum or spectra
}
\details{
Does not care or check if \code{colSums(sigs) == 1}.
Error checking is minimal since this function is called often.
}
\examples{
spectra <- PCAWG7::spectra$PCAWG$SBS96[, 1:2, drop = FALSE]
SBS96.sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
exposure <- PCAWG7::exposure$PCAWG$SBS96[, 1:2, drop = FALSE]
reconstructed.spectra <- ReconstructSpectrum(sigs = SBS96.sigs,
                                             exp = exposure,
                                             use.sig.names = TRUE)
}
