% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowSigActivity.R
\name{ShowExposure}
\alias{ShowExposure}
\title{Show signature exposure from the output generated by \code{AddSigActivity}}
\usage{
ShowExposure(
  list.of.sig.activity,
  output.dir,
  plot.exposure.proportion = FALSE
)
}
\arguments{
\item{list.of.sig.activity}{A list of contributing signature activity
information for multiple spectra. See the return value of
\code{\link{AddSigActivity}} for more details.}

\item{output.dir}{The directory to save the results. Create
this directory if it does not exist.}

\item{plot.exposure.proportion}{Whether to plot exposure proportions rather
than counts.}
}
\description{
Show signature exposure from the output generated by \code{AddSigActivity}
}
\keyword{internal}
