% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignaturePresenceTest.R
\name{SignaturePresenceTest}
\alias{SignaturePresenceTest}
\title{Test whether a given signature is plausibly present in a catalog.}
\usage{
SignaturePresenceTest(
  spectra,
  sigs,
  target.sig.index,
  m.opts = DefaultManyOpts(),
  seed = NULL,
  mc.cores = 2
)
}
\arguments{
\item{spectra}{The catalog (matrix) to analyze. This could be
an \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{sigs}{A catalog of signatures from which to choose.
This could be
and \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{target.sig.index}{The index of the signature the presence
of which we want to test. It can also be the signature id (e.g. "SBS22").}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}

\item{mc.cores}{Number of cores to use. Always silently
changed to 1 on Microsoft Windows.}
}
\value{
A list of test results for each sample in \code{spectra}.
Each sublist contains the following elements:
\itemize{
\item loglh.with: The maximum log likelihood of the reconstructed spectrum using
all the signatures.
\item loglh.without: The maximum log likelihood of the reconstructed spectrum
without the target signature.
\item statistic: Likelihood ratio test statistic.
\item chisq.p: P-value of the likelihood ratio test. The null hypothesis is we
can plausibly reconstruct the spectrum without the target signature.
\item exp.with: The exposure using all the signatures which generates the maximum
log likelihood \code{loglh.with}.
\item exp.without: The exposure not using the target signature which generates
the maximum log likelihood \code{loglh.without}.
}
}
\description{
Test whether a given signature is plausibly present in a catalog.
}
\examples{
indices <- grep("Lung-AdenoCA", colnames(PCAWG7::spectra$PCAWG$SBS96))
spectra <- PCAWG7::spectra$PCAWG$SBS96[, indices[1:2], drop = FALSE]
sigs <- PCAWG7::signature$genome$SBS96
sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                 cancer.type = "Lung-AdenoCA")
sigs.to.use <- sigs[, names(sigs.prop), drop = FALSE]
# Test whether SBS17a is plausibly present in the spectra
sig.presence.test.out <- SignaturePresenceTest(spectra = spectra,
                                               sigs = sigs.to.use, 
                                               target.sig.index = "SBS17a",
                                               seed = 2581,
                                               mc.cores = 2)
}
