% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SignaturePresenceTest}
\alias{SignaturePresenceTest}
\title{Test whether a given signature is plausibly present in a catalog}
\usage{
SignaturePresenceTest(
  spectra,
  sigs,
  target.sig.index,
  m.opts = NULL,
  mc.cores = 10
)
}
\arguments{
\item{spectra}{The catalog (matrix) to analyze. This could be
an \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{sigs}{A catalog of signatures from which to choose.
This could be
  and \code{\link[ICAMS]{ICAMS}} catalog or a numerical matrix.}

\item{target.sig.index}{The index of the signature the presence
of which we want to test.}

\item{m.opts}{If \code{NULL} use the return from
calling \code{\link{DefaultManyOpts}}. For documentation
see \code{\link{DefaultManyOpts}}.}

\item{mc.cores}{Number of cores to use. Always silently
changed to 1 on Microsoft Windows.}
}
\description{
Test whether a given signature is plausibly present in a catalog
}
