% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignActivity}
\alias{SparseAssignActivity}
\title{Find known signatures that can most sparsely reconstruct a spectrum.}
\usage{
SparseAssignActivity(spect, sigs, max.level = 5, p.thresh = 0.05,
  trace = 0, obj.fun, nbinom.size)
}
\arguments{
\item{spect}{The spectrum to be reconstructed.}

\item{sigs}{The signatures to use in reconstruction.}

\item{max.level}{The larges number of signatures to consider discarding
(this is not quite correct)}

\item{p.thresh}{The maximum p value based on which it is decided
to retain a signature in a reconstruction.}

\item{trace}{If > 0 print out information on progress.}

\item{obj.fun}{The objective function for \code{\link[nloptr]{nloptr}} ????.}

\item{nbinom.size}{The \code{size} parameter to \code{XXXXX}.}
}
\value{
An assignment of signature as ???.
}
\description{
Find known signatures that can most sparsely reconstruct a spectrum.
}
