% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignActivity}
\alias{SparseAssignActivity}
\title{Find known signatures that can most sparsely reconstruct each spectrum in a catalog.}
\usage{
SparseAssignActivity(
  spectra,
  sigs,
  max.level = 5,
  p.thresh = 0.05,
  eval_f = NULL,
  m.opts = NULL,
  mc.cores = NULL
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{The known signatures to use in reconstruction.}

\item{max.level}{The largest number of signatures to consider discarding
in the reconstruction.}

\item{p.thresh}{The maximum p value based on which it is decided
to retain a signature in a reconstruction.}

\item{eval_f}{The objective function for
\code{\link[nloptr]{nloptr}}.}

\item{m.opts}{For documentation
see \code{\link{DefaultManyOpts}}.}

\item{mc.cores}{The number of cores to use; if \code{NULL} use
\code{min(10, ncol(spectra))}, except on Windows, where \code{mc.cores}
is always 1.}
}
\value{
A list with the inferred exposure matrix as element \code{exposure}.
}
\description{
Find known signatures that can most sparsely reconstruct each spectrum in a catalog.
}
