% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseAssignActivity.R
\name{SparseAssignActivity}
\alias{SparseAssignActivity}
\title{Find known signatures that can most sparsely reconstruct each spectrum in a catalog.}
\usage{
SparseAssignActivity(
  spectra,
  sigs,
  max.level = 5,
  p.thresh = 0.05,
  m.opts = NULL,
  num.parallel.samples = 5,
  mc.cores.per.sample = min(20, 2^max.level)
)
}
\arguments{
\item{spectra}{The spectra (multiple spectra) to be reconstructed.}

\item{sigs}{The known signatures to use in reconstruction.}

\item{max.level}{The largest number of signatures to consider discarding
in the reconstruction.}

\item{p.thresh}{The maximum p value based on which it is decided
to retain a signature in a reconstruction.}

\item{m.opts}{For documentation
see \code{\link{DefaultManyOpts}}.}

\item{num.parallel.samples}{The (maximum) number of samples to run in parallel; each
sample in turn can require multiple cores, as governed by
\code{mc.cores.per.sample}.}

\item{mc.cores.per.sample}{The maximum number of cores to use for each sample.
On Microsoft Windows machines it is silently changed to 1.}
}
\value{
A list with the inferred exposure matrix as element \code{exposure}.
}
\description{
Find known signatures that can most sparsely reconstruct each spectrum in a catalog.
}
