% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{SparseAssignActivity1}
\alias{SparseAssignActivity1}
\title{Component of \code{\link{SparseAssignActivity}} for one spectrum.}
\usage{
SparseAssignActivity1(spect, sigs, max.level = 5, p.thresh = 0.05,
  trace = 0, eval_f, nbinom.size, global.opts, local.opts)
}
\description{
Component of \code{\link{SparseAssignActivity}} for one spectrum.
}
\keyword{internal}
