% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseAssignActivity1.R
\name{SparseAssignActivity1}
\alias{SparseAssignActivity1}
\title{Component of \code{\link{SparseAssignActivity}} for one spectrum.}
\usage{
SparseAssignActivity1(
  spect,
  sigs,
  max.level = 5,
  p.thresh = 0.05,
  m.opts,
  max.mc.cores = min(20, 2^max.level)
)
}
\arguments{
\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{The p value threshold for deciding if a set of signatures is necessary.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.)}
}
\description{
Component of \code{\link{SparseAssignActivity}} for one spectrum.
}
\keyword{internal}
