% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{TestSignaturePresenceTest1}
\alias{TestSignaturePresenceTest1}
\title{Framework for testing \code{\link{SignaturePresenceTest1}}.}
\usage{
TestSignaturePresenceTest1(sig.counts, trace = 0)
}
\description{
Framework for testing \code{\link{SignaturePresenceTest1}}.
}
\keyword{internal}
