% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSigAct.R
\name{TestSignaturePresenceTest1}
\alias{TestSignaturePresenceTest1}
\title{Framework for testing \code{\link{SignaturePresenceTest1}}.}
\usage{
TestSignaturePresenceTest1(
  sig.counts,
  input.sigs = PCAWG7::signature$genome$SBS96,
  trace = 0,
  eval_f = ObjFnBinomMaxLHNoRoundOK,
  m.opts = NULL
)
}
\description{
Framework for testing \code{\link{SignaturePresenceTest1}}.
}
\keyword{internal}
