% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSparseAssignOneSample.R,
%   R/SparseAssignActivity1.R
\name{tryingtoremove}
\alias{tryingtoremove}
\title{Run \code{SparseAssignActivity1} on one sample and save and plot the results}
\arguments{
\item{output.dir}{Directory path to save the output file.}

\item{spect}{A single spectrum.}

\item{sigs}{A numerical matrix, possibly an \code{\link[ICAMS]{ICAMS}} catalog.}

\item{max.level}{The maximum number of signatures to try removing.}

\item{p.thresh}{The p value threshold for deciding if a set of signatures is necessary.}

\item{m.opts}{See \code{\link{DefaultManyOpts}}.}

\item{max.mc.cores}{The maximum number of cores to use.
On Microsoft Windows machines it is silently changed to 1.)}

\item{seed}{Random seed; set this to get reproducible results. (The
numerical optimization is in two phases; the first, global phase
might rarely find different optima depending on the random
seed.)}
}
\description{
Run \code{SparseAssignActivity1} on one sample and save and plot the results

Component of \code{\link{SparseAssignActivity}} for one spectrum.
}
\keyword{internal}
